==> analysis/0_csvToDataframe_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd
import os

from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta, SU, MO, TU, WE, TH, FR, SA

## df processing

all_df = []

for filename in os.listdir('Data/message_csv/'):
    if filename == '.DS_Store':
        continue
    print(filename)
    df = pd.read_csv('Data/message_csv/' + filename)
    df = df.drop(df.columns[0], axis=1)

    all_df.append(df)
    
df_x = pd.concat(all_df).reset_index(drop = True)

# Some problematic message times
print(df_x.shape)
df_x = df_x.drop(np.where(df_x['message_time'] == 'View')[0], axis = 0)
print(df_x.shape)

df_x = df_x.reset_index(drop = True)

print(df_x[df_x['message_time'] == 'View'].shape)

# Drop some error dates
print(df_x.shape)
df_x = df_x.drop(np.where(df_x['date'].isnull())[0], axis = 0)
print(df_x.shape)

df_x = df_x.reset_index(drop = True)

print(df_x[df_x['date'].isnull()].shape)

df_x["message_time"] = pd.to_datetime(df_x["message_time"], format='%I:%M %p').dt.time

df_x["mtime"] = pd.to_datetime(df_x["mtime"], format='%Y-%m-%d %H:%M:%S')
df_x["timez"] = pd.to_datetime(df_x["timez"], format='%Y-%m-%d_%H%M%SZ')

df_x.loc[df_x['server'] == 'CALIFORNIA', 'mtime'] += timedelta(hours = 6)

df_x.loc[df_x['mtime'] < df_x['timez'], ['mtime']] += timedelta(hours = 1)

df_x.loc[df_x['mtime'] < df_x['timez']]

# Only save telephone number from message origin

df_x["tel"] = df_x['message_origin'].str.split("\'").str[1]
df_x = df_x.drop('message_origin', axis = 1)

# Find rows where start_traverse and this_traverse date differ

# i.e., if I start traversing at 23:55, and my read time on messages is at 00:01,
# which day are WhatsApp labels in reference to?

# this is only a problem for the FIRST GROUP READ with mismatch dates for each server/read-time pair
# all groups after are guaranteed to have been read on the second day

i_mismatch = df_x['timez'].dt.date != df_x['mtime'].dt.date
mismatch = df_x[i_mismatch].sort_values(['mtime'], ascending=True).groupby(['timez', 'server']).head(1)[['timez', 'server', 'uid']]

mismatch = mismatch.values.tolist()

def gen_message_date(row):
    if row['date'] != 'TODAY' and row['date'] != 'YESTERDAY':
        if '/' in row['date']:
            return datetime.strptime(row['date'], '%m/%d/%Y').date()

        elif row['date'] == 'SUNDAY':
            return row['mtime'].date() + relativedelta(weekday=SU(-1))

        elif row['date'] == 'MONDAY':
            return row['mtime'].date() + relativedelta(weekday=MO(-1))

        elif row['date'] == 'TUESDAY':
            return row['mtime'].date() + relativedelta(weekday=TU(-1))

        elif row['date'] == 'WEDNESDAY':
            return row['mtime'].date() + relativedelta(weekday=WE(-1))

        elif row['date'] == 'THURSDAY':
            return row['mtime'].date() + relativedelta(weekday=TH(-1))

        elif row['date'] == 'FRIDAY':
            return row['mtime'].date() + relativedelta(weekday=FR(-1))

        elif row['date'] == 'SATURDAY':
            return row['mtime'].date() + relativedelta(weekday=SA(-1))
    
    elif row['timez'].date() == row['mtime'].date():
        cur_date = row['mtime'].date()
        if row['date'] == 'TODAY':
            return cur_date
        elif row['date'] == 'YESTERDAY':
            return cur_date - timedelta(days = 1)
        
    else:
        if [row['timez'], row['server'], row['uid']] in mismatch:
            return 'FLAG'
        else:
            cur_date = row['mtime'].date()
            if row['date'] == 'TODAY':
                return cur_date
            elif row['date'] == 'YESTERDAY':
                return cur_date - timedelta(days = 1)

def mdate_message_date(row):    
    cur_date = row['mtime'].date()
    if row['date'] == 'TODAY':
        return cur_date
    elif row['date'] == 'YESTERDAY':
        return cur_date - timedelta(days = 1)

df_x['message_date'] = df_x.apply(lambda x: gen_message_date(x), axis = 1)

# Set all first based on infocheck

for index, row in df_x[df_x['message_date'] == 'FLAG'].iterrows():
    if pd.notnull(row['infocheck']):
        a = row['infocheck'].find(',') + 2
        b = row['infocheck'].find(']')
        
        infocheck_date = datetime.strptime(row['infocheck'][a:b], '%m/%d/%Y').date()
        df_x.loc[index, 'message_date'] = infocheck_date

# Then set to mtime

df_x.loc[df_x['message_date'] == 'FLAG', 'message_date'] =\
    df_x.loc[df_x['message_date'] == 'FLAG'].apply(lambda x: mdate_message_date(x), axis = 1)

df_x = df_x.drop(['date', 'infocheck'], axis = 1)

# drop information
df_x = df_x[df_x['message_time'].notnull()]
df_x['message_dt'] = pd.to_datetime(df_x['message_date'].astype(str) + ' ' + df_x['message_time'].astype(str))

df_x = df_x.rename_axis('idx')
df_x = df_x.sort_values(by=['message_dt', 'idx'])

## Removing duplicates

We want to remove all duplicated messages, since it's possible that our traverseScripts reads the same message on different runs (in particular, if there are few new messages betwen two runs of the script, it will encounter many of the same messages, since that's what WhatsApp loads).

However, we don't want to remove bona fide duplicates, that is, duplicate messages that were actually sent twice. We can identify bona fide duplicates as duplicates with the same timez (referring to the time they were scraped).

# Duplicate messages within same traverse
intra_traverse_duplicates = df_x[df_x.duplicated(keep = False)]

# This gives us all duplicates that have the same timez.
# However, bona fide duplicates may be duplicated across multiple runs of traverse, so we only
# want one appearance of them.

# We can go through all possible bona fide duplicates
# and mark to keep them iff they are the first run bona fide duplicate

duplicates_first_appearance = []
i_true_duplicates = []
partials_collection = []

for i, row in intra_traverse_duplicates.iterrows():
    row_partial = row[['uid', 'event', 'text',
       'message_time', 'infocheck', 'emojis', 'audio_length', 'image',
       'reply_to', 'reply_text', 'reply_image', 'reply_thumb', 'reply_emojis',
       'reply_type_audio', 'reply_type_image', 'message_deleted',
       'video_length', 'video_thumb', 'forwarded', 'forwarded_highly', 'tel',
       'message_date']]

    if str(row.values) in duplicates_first_appearance:
        i_true_duplicates.append(i)        
    else:
        if str(row_partial.values) in partials_collection:
            pass
        else:
            duplicates_first_appearance.append(str(row.values))
            partials_collection.append(str(row_partial.values))
            
print(len(i_true_duplicates))

persist_cols = ['uid', 'event', 'text',
       'message_time', 'emojis', 'audio_length', 'image', 'reply_to',
       'reply_text', 'reply_image', 'reply_thumb', 'reply_emojis',
       'reply_type_audio', 'reply_type_image', 'message_deleted',
       'video_length', 'video_thumb', 'forwarded', 'forwarded_highly', 'tel',
       'message_date']

i_duplicates = df_x[df_x[persist_cols].duplicated(keep = 'first')].index.values

i_duplicates_remove = [i for i in i_duplicates if i not in i_true_duplicates]

print(len(i_duplicates))
print(len(i_duplicates_remove))

df_x = df_x.drop(i_duplicates_remove, axis = 0)

# Reset dates times based on UTC-5
df_x['message_dt'] -= timedelta(hours = 5)
df_x = df_x.drop(['message_date', 'message_time'], axis = 1)

df_x['message_date'] = df_x['message_dt'].dt.date
df_x = df_x.reset_index(drop = True)

print(df_x.shape)

df_x = df_x[df_x.message_date >= pd.Timestamp(2020, 2, 13)]
df_x = df_x[df_x.message_date <= pd.Timestamp(2020, 4, 5)]

print(df_x.shape)

df_x.to_csv("Data/dfx_nb0.csv")
