==> analysis/0a_groupMerge_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd

import string, re, unidecode

import nltk
from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer

from collections import Counter

from sklearn.feature_extraction.text import TfidfVectorizer

from sklearn.metrics.pairwise import cosine_similarity

# Read in

df_x = pd.read_csv("Data/dfx_nb0.csv", index_col = 0)
print(df_x.shape)

df_x['text'] = df_x['text'].fillna('')
df_x['textlower'] = df_x['text'].str.lower()
df_x['textlower'] = df_x['textlower'].str.replace('\n', ' ') # Replace all newlines

# Just for bottom print
df_x['title'] = df_x['title'].fillna('')

# Text processing

stemmer = SnowballStemmer('spanish')
remove_punc = str.maketrans('', '', string.punctuation)
stopwords_ascii = [unidecode.unidecode(w) for w in stopwords.words('spanish')]

def tokenize(s):
    s = unidecode.unidecode(s)
    s = s.translate(remove_punc)
    tokens = nltk.word_tokenize(s)
    filtered = [w for w in tokens if w not in stopwords_ascii]
    return [stemmer.stem(w) for w in filtered]

def dummy(x):
    return x

vectorizer = TfidfVectorizer(analyzer='word', tokenizer=dummy, preprocessor=dummy,
                            token_pattern=None)

df_t = df_x[df_x['textlower'].str.len() > 0]
print(df_t.shape)

## Tokenize

df_t.loc[:,'token'] = df_t.loc[:,'textlower'].apply(tokenize)

# Tokenize across all groups
def aggTokens(df):
    return df['token'].sum()

df_tgroup = df_t.groupby('uid').apply(aggTokens)

groups = sorted(df_t.uid.unique())
groups_titles = [df_x[df_x['uid'] == g]['title'].unique() for g in groups]

X = vectorizer.fit_transform(df_tgroup)
c = cosine_similarity(X)

def groupMsg(i):
    return df_x[df_x['uid'] == groups[i]]

def groupLatestDate(i):
    return groupMsg(i).sort_values('message_dt').iloc[-1]['message_dt']

trueCols = ['event', 'text', 'emojis',
       'audio_length', 'image', 'reply_to', 'reply_text', 'reply_image',
       'reply_thumb', 'reply_emojis', 'reply_type_audio', 'reply_type_image',
       'message_deleted', 'video_length', 'video_thumb', 'forwarded',
       'forwarded_highly', 'tel', 'message_dt', 'message_date', 'textlower']

## Process similarities

Match if one group contains 60\% of the other's messages (full inner merge)

rename_from_to = []

for i in zip(np.nonzero(c > 0.5)[0], np.nonzero(c > 0.5)[1]):
    if (i[0] >= i[1]):
        continue
    print(c[i[0], i[1]])
    print(i)
    print(groups[i[0]] + " / " + groups[i[1]])
    print(groups_titles[i[0]])
    print(groups_titles[i[1]])
    merge_size = pd.merge(groupMsg(i[0])[trueCols], groupMsg(i[1])[trueCols], how='inner').shape[0]
    
    if merge_size > 0.6 * min(groupMsg(i[0])[trueCols].shape[0], groupMsg(i[1])[trueCols].shape[0]) \
        and not (('-' in groups[i[0]]) and ('-' in groups[i[1]])):
        print("\nMERGE")
        
        if '-' in groups[i[0]]:
            rename_from_to.append((groups[i[1]], groups[i[0]]))
        elif '-' in groups[i[1]]:
            rename_from_to.append((groups[i[0]], groups[i[1]]))
        elif groupLatestDate(i[0]) >= groupLatestDate(i[1]):
            rename_from_to.append((groups[i[1]], groups[i[0]]))
        else:
            rename_from_to.append((groups[i[0]], groups[i[1]]))
    
    else:
        print()
        print(merge_size / min(groupMsg(i[0])[trueCols].shape[0], groupMsg(i[1])[trueCols].shape[0]))
        
    print()

print(len(rename_from_to))
print(df_x['uid'].unique().shape[0])

with open('Data/intermediate/rename_from_to.txt', 'w') as file:
    file.write(str(rename_from_to))

## Fix

df_x['uid'] = df_x['uid'].replace([i[0] for i in rename_from_to], [i[1] for i in rename_from_to])

# Remove duplicates again

# Duplicate messages within same traverse
intra_traverse_duplicates = df_x[df_x.duplicated(keep = False)]

# This gives us all duplicates that have the same timez.
# However, bona fide duplicates may be duplicated across multiple runs of traverse, so we only
# want one appearance of them.

# We can go through all possible bona fide duplicates
# and mark to keep them iff they are the first run bona fide duplicate

duplicates_first_appearance = []
i_true_duplicates = []
partials_collection = []

for i, row in intra_traverse_duplicates.iterrows():
    row_partial = row[['uid', 'event', 'text', 'emojis',
       'audio_length', 'image', 'reply_to', 'reply_text', 'reply_image',
       'reply_thumb', 'reply_emojis', 'reply_type_audio', 'reply_type_image',
       'message_deleted', 'video_length', 'video_thumb', 'forwarded',
       'forwarded_highly', 'tel', 'message_dt', 'message_date', 'textlower']]

    if str(row.values) in duplicates_first_appearance:
        i_true_duplicates.append(i)        
    else:
        if str(row_partial.values) in partials_collection:
            pass
        else:
            duplicates_first_appearance.append(str(row.values))
            partials_collection.append(str(row_partial.values))
            
print(len(i_true_duplicates))

persist_cols = ['uid', 'title', 'event', 'text', 'emojis',
       'audio_length', 'image', 'reply_to', 'reply_text', 'reply_image',
       'reply_thumb', 'reply_emojis', 'reply_type_audio', 'reply_type_image',
       'message_deleted', 'video_length', 'video_thumb', 'forwarded',
       'forwarded_highly', 'tel', 'message_dt', 'message_date', 'textlower']

i_duplicates = df_x[df_x[persist_cols].duplicated(keep = 'first')].index.values

i_duplicates_remove = [i for i in i_duplicates if i not in i_true_duplicates]

print(len(i_duplicates))
print(len(i_duplicates_remove))

df_x = df_x.drop(i_duplicates_remove, axis = 0)
print(df_x.shape)

df_x = df_x.reset_index(drop = True)

df_x.to_csv("Data/df_x_nb0a-groupMerge.csv")
