==> analysis/1_members_byMessaging_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd
import os

from collections import Counter

import matplotlib.pyplot as plt
SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

if not os.path.exists('images/ch-members/active/'):
    os.makedirs('images/ch-members/active/')

import statsmodels.formula.api as smf
import scipy.stats
from scipy.stats import ttest_ind
from scipy.stats import f_oneway

import networkx as nx

# Read in

df_x = pd.read_csv("Data/df_x_nb0a-groupMerge.csv", index_col = 0)
print(df_x.shape)

### Skip this processing

group_members = df_x[['uid', 'tel']].groupby('uid')['tel'].unique()

df_interact = pd.DataFrame()
df_groups = pd.DataFrame()

for member_list in group_members:
    member_list = list(member_list)
    df_new = pd.DataFrame(index = member_list, columns = member_list + ['Group']).fillna(1)
        
    df_interact = df_interact.add(df_new, fill_value = 0)

    country_codes = [m.split(' ')[0] for m in member_list]
    df_groups = df_groups.append(Counter(country_codes), ignore_index=True)

df_groups.index = group_members.index
df_groups = df_groups.fillna(0)
df_interact = df_interact.fillna(0)

df_groups.to_csv('Data/intermediate/df_groups_nb1-members-byMessaging.csv')
df_interact.to_csv('Data/intermediate/df_interact_nb1-members-byMessaging.csv')

### Reload from saved output

# Re-import from first run
df_groups = pd.read_csv('Data/intermediate/df_groups_nb1-members-byMessaging.csv', index_col = 0)
df_interact = pd.read_csv('Data/intermediate/df_interact_nb1-members-byMessaging.csv', index_col = 0)

# Group size

print("Group/interact shapes:")
print(df_groups.shape)
print(df_interact.shape)

count_cols = [c for c in df_groups.columns if c[0] == '+']
df_groups["Size"] = df_groups[count_cols].sum(axis = 1)

print("Group sizes <=10, <=20, >=150")
print(np.mean(df_groups['Size'] <= 10))
print(np.mean(df_groups['Size'] <= 20))
print(np.mean(df_groups['Size'] >= 150))

# distribution of group sizes

plt.hist(df_groups["Size"], range = (0,256), bins = 20)
plt.xlabel("# of Members in Group")
plt.ylabel("# of Groups")
plt.savefig('images/ch-members/active/hist_size.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# distribution of groups/user
print("Groups/user value counts:")
print(df_interact["Group"].value_counts())

# Country breakdown

df_groups["pVZ"] = df_groups["+58"] / df_groups["Size"]
df_groups["pCO"] = df_groups["+57"] / df_groups["Size"]
df_groups["pUS"] = df_groups["+1"] / df_groups["Size"]
df_groups["pPE"] = df_groups["+51"] / df_groups["Size"]
df_groups["pCL"] = df_groups["+56"] / df_groups["Size"]
df_groups["pEC"] = df_groups["+593"] / df_groups["Size"]
df_groups['p3rdCountry'] = (1 - (df_groups['pVZ'] + df_groups['pCO']))

# distribution of pCountry in various groups
# APPENDIX

for col in ["pVZ", "pCO", "pUS", "pPE", "pCL", "pEC", "p3rdCountry"]:
    plt.hist(100 * df_groups[col], range = (0,100), bins = 16)
    plt.xlabel("%% of Group Members from %s" % col[1:])
    plt.ylabel("# of Groups")   
    plt.savefig('images/ch-appendicies/members/prop_%s.png' % col[1:], bbox_inches = 'tight', pad_inches = 0.05)
    plt.close()

# comparing VZ and CO membership in groups

plt.scatter(100 * df_groups["pVZ"], 100 * df_groups["pCO"], alpha = 0.5)
plt.xlabel("% VZ in Group")
plt.ylabel("% CO in Group")

plt.savefig('images/ch-members/active/pVZ_vs_pCO.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.hist(df_groups[df_groups["pVZ"] > df_groups["pCO"]]["Size"], range = (0,256), bins = 20)
plt.xlabel("Size of Group")
plt.ylabel("Number of Groups")
plt.title("Groups where %VZ > %CO")
plt.savefig('images/ch-members/active/hist_size_moreVZ.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.hist(df_groups[df_groups["pVZ"] < df_groups["pCO"]]["Size"], range = (0,256), bins = 20)
plt.xlabel("Size of Group")
plt.ylabel("Number of Groups")
plt.title("Groups where %VZ < %CO")
plt.savefig('images/ch-members/active/hist_size_moreCO.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

## Who's connected to who

df_interact["CO Connections"] = df_interact[[col for col in df_interact if col[:4] == "+57 "]].sum(axis = 1)
df_interact["VZ Connections"] = df_interact[[col for col in df_interact if col[:4] == "+58 "]].sum(axis = 1)
df_interact["PE Connections"] = df_interact[[col for col in df_interact if col[:4] == "+51 "]].sum(axis = 1)
df_interact["US Connections"] = df_interact[[col for col in df_interact if col[:3] == "+1 "]].sum(axis = 1)
df_interact["BR Connections"] = df_interact[[col for col in df_interact if col[:4] == "+55 "]].sum(axis = 1)
df_interact["CL Connections"] = df_interact[[col for col in df_interact if col[:4] == "+56 "]].sum(axis = 1)
df_interact["EC Connections"] = df_interact[[col for col in df_interact if col[:5] == "+593 "]].sum(axis = 1)
df_interact["nConnections"] = df_interact[[col for col in df_interact if col[0] == "+"]].sum(axis = 1)

df_interact_vz = df_interact.loc[[num for num in df_interact.index if num[:4] == "+58 "]]
df_interact_co = df_interact.loc[[num for num in df_interact.index if num[:4] == "+57 "]]
df_interact_pe = df_interact.loc[[num for num in df_interact.index if num[:4] == "+51 "]]
df_interact_cl = df_interact.loc[[num for num in df_interact.index if num[:4] == "+56 "]]
df_interact_ec = df_interact.loc[[num for num in df_interact.index if num[:5] == "+593 "]]
df_interact_us = df_interact.loc[[num for num in df_interact.index if num[:3] == "+1 "]]

### Comparing connections to VZ and CO users


plt.hist(df_interact_vz["CO Connections"], label = "To CO Users", range = (0,300), alpha = 0.5, bins = 30)
plt.hist(df_interact_vz["VZ Connections"], label = "To VZ Users", range = (0,300), alpha = 0.5, bins = 30)
plt.xlabel("Number of Connections")
plt.ylabel("Number of Users")
plt.title("Venezuelan WhatsApp Users (+58)")
plt.legend()
plt.savefig("images/ch-members/active/hist_connToCOVZ_VZ.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.hist(df_interact_co["CO Connections"], label = "To CO Users", range = (0,300), alpha = 0.5, bins = 30)
plt.hist(df_interact_co["VZ Connections"], label = "To VZ Users", range = (0,300), alpha = 0.5, bins = 30)
plt.xlabel("Number of Connections")
plt.ylabel("Number of Users")
plt.title("Colombian WhatsApp Users (+57)")
plt.legend()
plt.savefig("images/ch-members/active/hist_connToCOVZ_CO.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.hist(df_interact_pe["CO Connections"], label = "To CO Users", range = (0,300), alpha = 0.5, bins = 30)
plt.hist(df_interact_pe["VZ Connections"], label = "To VZ Users", range = (0,300), alpha = 0.5, bins = 30)
plt.xlabel("Number of Connections")
plt.ylabel("Number of Users")
plt.title("Peruvian WhatsApp Users (+51)")
plt.legend()
plt.savefig("images/ch-members/active/hist_connToCOVZ_PE.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.hist(df_interact_ec["CO Connections"], label = "To CO Users", range = (0,300), alpha = 0.5, bins = 30)
plt.hist(df_interact_ec["VZ Connections"], label = "To VZ Users", range = (0,300), alpha = 0.5, bins = 30)
plt.xlabel("Number of Connections")
plt.ylabel("Number of Users")
plt.title("Ecuadorian WhatsApp Users (+593)")
plt.legend()
plt.savefig("images/ch-members/active/hist_connToCOVZ_EC.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.hist(df_interact_cl["CO Connections"], label = "To CO Users", range = (0,300), alpha = 0.5, bins = 30)
plt.hist(df_interact_cl["VZ Connections"], label = "To VZ Users", range = (0,300), alpha = 0.5, bins = 30)
plt.xlabel("Number of Connections")
plt.ylabel("Number of Users")
plt.title("Chilean WhatsApp Users (+56)")
plt.legend()
plt.savefig("images/ch-members/active/hist_connToCOVZ_CL.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Bar graphs of user connected to users from each country

def selectCountryCode(df, cc):
    return df[df.index.str.startswith(cc + ' ')]

def allCountryCode(df):
    return [x[0] for x in df.index.str.split(" ")]

cc = {'+51': 'PER', '+55': 'BRA', '+56': 'CHL', '+57': 'COL', '+58': 'VEN',
        '+593': 'ECU', '+1': 'USA'}

def countryCodeToCountry(code):
    return cc[code]

selectCountries = cc.keys()

# bar graph by country

df_interest = df_interact
countCountryCode = Counter(allCountryCode(df_interest))

x = [countryCodeToCountry(c) for c in countCountryCode and selectCountries]
y = [100 * countCountryCode[c] / len(df_interest) for c in countCountryCode and selectCountries]

plt.figure(figsize = (9, 5))
plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("% of Users")
plt.title("All Users (%s)" % len(df_interest))
plt.savefig("images/ch-members/active/bar_byCountry.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# bar graph by country, for those with Peru connections

df_interest = df_interact[df_interact["PE Connections"] > 0]
countCountryCode = Counter(allCountryCode(df_interest))

x = [countryCodeToCountry(c) for c in countCountryCode and selectCountries]
y = [100 * countCountryCode[c] / len(df_interest) for c in countCountryCode and selectCountries]

plt.figure(figsize = (9, 5))
plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("% of Users")
plt.title("Users Connected to Peru Users (%s)" % len(df_interest))
plt.savefig("images/ch-members/active/bar_withConnTo_PE.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# bar graph by country, for those with Chile connections

df_interest = df_interact[df_interact["CL Connections"] > 0]
countCountryCode = Counter(allCountryCode(df_interest))

x = [countryCodeToCountry(c) for c in countCountryCode and selectCountries]
y = [100 * countCountryCode[c] / len(df_interest) for c in countCountryCode and selectCountries]

plt.figure(figsize = (9, 5))
plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("% of Users")
plt.title("Users Connected to Chile Users (%s)" % len(df_interest))
plt.savefig("images/ch-members/active/bar_withConnTo_CL.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

""# bar graph by country, for those with Ecuador connections

df_interest = df_interact[df_interact["EC Connections"] > 0]
countCountryCode = Counter(allCountryCode(df_interest))

x = [countryCodeToCountry(c) for c in countCountryCode and selectCountries]
y = [100 * countCountryCode[c] / len(df_interest) for c in countCountryCode and selectCountries]

plt.figure(figsize = (9, 5))
plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("% of Users")
plt.title("Users Connected to Ecuador Users (%s)" % len(df_interest))
plt.savefig("images/ch-members/active/bar_withConnTo_EC.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# Diversity

### Entropy

def entr(row):
    return scipy.stats.entropy(row[count_cols])

df_groups['entropy'] = df_groups.apply(entr, axis = 1)

# distribution of group entropies

plt.hist(df_groups["entropy"], bins = 15)
plt.xlabel("Entropy")
plt.ylabel("# of Groups")
plt.title("Diversity within Groups")
plt.savefig('images/ch-members/active/hist_entropy.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Simpson

def simpson(row):
    counts = row[count_cols]
    return np.sum(np.square(counts)) / (row['Size']**2)

df_groups['simpson'] = df_groups.apply(simpson, axis = 1)

### Entropy vs. Simpson

plt.scatter(df_groups['simpson'], df_groups['entropy'])
plt.xlabel("Simpson Index (similarity)")
plt.ylabel("Entropy (uncertainty)")
plt.title("National Diversity within Groups")
plt.savefig('images/ch-members/active/scatter_simpson_entropy.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print("Entropy ~ simpson")

reg = smf.ols('entropy ~ simpson', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['entropy'], df_groups['simpson']))
print(reg.summary())

## Diversity correlates

plt.scatter(df_groups['pVZ'], df_groups['entropy'])
plt.xlabel("Proportion VZ in Group")
plt.ylabel("Entropy")
plt.savefig('images/ch-members/active/scatter_pVZ_entropy.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.scatter(df_groups['pCO'], df_groups['entropy'])
plt.xlabel("Proportion CO in Group")
plt.ylabel("Entropy")
plt.savefig('images/ch-members/active/scatter_pCO_entropy.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print("Entropy ~ pVZ")

reg = smf.ols('entropy ~ pVZ', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['pVZ'], df_groups['entropy']))
print(reg.summary())

print("Entropy ~ pCO")
reg = smf.ols('entropy ~ pCO', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['pCO'], df_groups['entropy']))
print(reg.summary())

plt.scatter(df_groups['p3rdCountry'], df_groups['entropy'])
plt.xlabel("Prop. Non-CO/Non-VZ in Group")
plt.ylabel("Entropy")
plt.savefig('images/ch-members/active/scatter_p3rd_entropy.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print("Entropy ~ p3rd")

reg = smf.ols('entropy ~ p3rdCountry', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['p3rdCountry'], df_groups['entropy']))
print(reg.summary())

plt.scatter(df_groups['Size'], df_groups['entropy'])
plt.xlabel("Group Size")
plt.ylabel("Entropy")
plt.savefig('images/ch-members/active/scatter_size_entropy.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print("Entropy ~ Size")

reg = smf.ols('entropy ~ Size', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['Size'], df_groups['entropy']))
print(reg.summary())

### Simpson, to check

print("Simpson ~ pVZ, pCO, p3rd, Size (to check)")
reg = smf.ols('simpson ~ pVZ', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['pVZ'], df_groups['simpson']))
print(reg.summary())

reg = smf.ols('simpson ~ pCO', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['pCO'], df_groups['simpson']))
print(reg.summary())

reg = smf.ols('simpson ~ p3rdCountry', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['p3rdCountry'], df_groups['simpson']))
print(reg.summary())

reg = smf.ols('simpson ~ Size', data = df_groups).fit()
print(scipy.stats.pearsonr(df_groups['Size'], df_groups['simpson']))
print(reg.summary())

# Network: Groups as nodes

members_many_groups = df_interact.index[df_interact['Group'] > 1]

df_interact_groups_nodes = pd.DataFrame()

for member in members_many_groups:
    group_list = list(df_x[df_x['tel'] == member]['uid'].unique())
    df_new = pd.DataFrame(index = group_list, columns = group_list).fillna(1)
        
    df_interact_groups_nodes = df_interact_groups_nodes.add(df_new, fill_value = 0)

df_interact_groups_nodes = df_interact_groups_nodes.fillna(0)

matr_interact = df_interact_groups_nodes.values

df_groups['degree'] = df_interact_groups_nodes.sum(axis = 1)
df_groups['degree'] = df_groups['degree'].fillna(0)

G = nx.from_numpy_matrix(matr_interact)

Gcc = sorted(nx.connected_components(G), key=len, reverse=True)
G0 = G.subgraph(Gcc[0])

average_dists = []

for i in list(G):
    lengths = nx.shortest_path_length(G, source = i).values()
    average_dists.append(sum(lengths) / (len(lengths) - 1))

print("Group graph")
print("Largest CC n_nodes: %s" % G0.number_of_nodes())
print("Graph n_nodes: %s" % G.number_of_nodes())

### Groups networkx graph

nodes_co = [i for i in list(G) if df_groups.loc[df_interact_groups_nodes.index[i], 'pCO'] >= 0.5]
nodes_vz = [i for i in list(G) if df_groups.loc[df_interact_groups_nodes.index[i], 'pVZ'] >= 0.5]
nodes_3rd = [i for i in list(G) if i not in nodes_co and i not in nodes_vz]

pos = nx.spring_layout(G, k = 1.2, iterations = 200, seed = 35)
plt.figure(figsize = (14, 8))

nx.draw_networkx_nodes(G, pos=pos, nodelist = nodes_co, node_size = 50, node_color = 'blue', alpha = 0.7)
nx.draw_networkx_nodes(G, pos=pos, nodelist = nodes_vz, node_size = 50, node_color = 'red', alpha = 0.7)
nx.draw_networkx_nodes(G, pos=pos, nodelist = nodes_3rd, node_size = 50, node_color = 'orange', alpha = 0.7)
nx.draw_networkx_nodes(G, pos=pos, nodelist = [nodes_co[0]], node_size = 50, node_color = 'blue', alpha = 0.7, label = '>50% CO')
nx.draw_networkx_nodes(G, pos=pos, nodelist = [nodes_vz[0]], node_size = 50, node_color = 'red', alpha = 0.7, label = '>50% VZ')
nx.draw_networkx_nodes(G, pos=pos, nodelist = [nodes_3rd[0]], node_size = 50, node_color = 'orange', alpha = 0.7, label = 'Neither')

nx.draw_networkx_edges(G, pos=pos, width = 1)
plt.legend()
plt.title("Groups as Nodes (connected if they share members)")
plt.savefig('images/ch-members/active/networkx_groups_co_vz_3rd.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

dists_co = [average_dists[i] for i in nodes_co if i in Gcc[0]]
dists_vz = [average_dists[i] for i in nodes_vz if i in Gcc[0]]
dists_3rd = [average_dists[i] for i in nodes_3rd if i in Gcc[0]]

print('CO Centrality: %.3f' % np.mean(dists_co))
print('VZ Centrality: %.3f' % np.mean(dists_vz))
print('3rd Centrality: %.3f' % np.mean(dists_3rd))

print(f_oneway(dists_co, dists_vz, dists_3rd))
print(ttest_ind(dists_co, dists_vz))

print("Group graph, average dist <= 3: number and number VZ")
co = 0
vz = 0
print(len([i for i in Gcc[0] if average_dists[i] <= 3]))
print(len([i for i in nodes_vz if average_dists[i] <= 3 and i in Gcc[0]]))


## Correlates of centrality

df_interact_groups_nodes['avg_dist'] = average_dists

df_groups['G0_avg_dist'] = df_interact_groups_nodes.iloc[list(G0)]['avg_dist']
df_groups_G0 = df_groups[df_groups['G0_avg_dist'].notnull()]

print("With G0_avg_dist")
for col in ['Size', 'pVZ', 'pCO', 'pUS', 'pPE', \
       'pCL', 'pEC', 'p3rdCountry', 'entropy', 'simpson']:
    
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups_G0[col], df_groups_G0['G0_avg_dist']))

print("G0_avg_dist ~ Size")
reg = smf.ols('G0_avg_dist ~ Size', data = df_groups).fit()
print(reg.summary())

## Correlates of group degree

# distribution of group degrees

plt.hist(df_groups["degree"], bins = 30)
plt.xlabel("Degree")
plt.ylabel("# of Groups")
plt.title("Group Network")
plt.savefig('images/ch-members/active/hist_degree.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print("Group degree: mean, median, <5 prop, >20 prop")
print(np.mean(df_groups["degree"]))
print(np.median(df_groups["degree"]))
print(np.mean(df_groups["degree"] < 5))
print(np.mean(df_groups["degree"] > 20))

print("With degree")

for col in ['Size', 'pVZ', 'pCO', 'pUS', 'pPE', \
       'pCL', 'pEC', 'p3rdCountry', 'entropy', 'simpson']:
    
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[col], df_groups['degree']))
    
print("============= G0_avg_dist =============")
print(scipy.stats.pearsonr(df_groups_G0['degree'], df_groups_G0['G0_avg_dist']))

print("Degree ~ size + entropy")

reg = smf.ols('degree ~ Size + entropy', data = df_groups).fit()
print(reg.summary())

print("Degree ~ size + entropy, only degree <= 70")

reg = smf.ols('degree ~ Size + entropy', data = df_groups[df_groups['degree'] <= 70]).fit()
print(reg.summary())

## Clustering coefficient

groups_clustering = pd.DataFrame(index = df_interact_groups_nodes.index.values,
                                 data = nx.clustering(G).values(), columns = ['clustering'])

df_groups['clustering'] = groups_clustering['clustering']

df_c = df_groups[df_groups['clustering'].notna()]

print("df_c (clustering) shape")
print(df_c.shape)

# distribution of group sizes

plt.hist(df_c["clustering"], bins = 20)
plt.xlabel("Clustering Coefficient")
plt.ylabel("# of Groups")
plt.title("Group Network")
plt.savefig('images/ch-members/active/hist_clustering.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print("With clustering")

for col in ['Size', 'pVZ', 'pCO', 'pUS', 'pPE', \
       'pCL', 'pEC', 'p3rdCountry', 'entropy', 'simpson', 'degree']:
    
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_c[col], df_c['clustering']))

## 3 Class graphs (groups)

def threeClassGraphG0(colname, label, appendix = False):
    def valsMap(uid):
        return df_groups.loc[uid, colname]
    
    l = np.quantile(df_groups[colname], .3)
    h = np.quantile(df_groups[colname], .7)

    vals = [valsMap(u) for u in df_interact_groups_nodes.index.values]

    nodes_low = [i for i in Gcc[0] if vals[i] <= l]
    nodes_mid = [i for i in Gcc[0] if vals[i] > l and vals[i] <= h]
    nodes_high = [i for i in Gcc[0] if vals[i] > h]
    
    plt.figure(figsize = (14, 14))
    pos = nx.spring_layout(G0, iterations = 200, seed = 0)
    nx.draw_networkx_nodes(G0, pos=pos, nodelist=nodes_low, node_color='#ffb1b1',
                           label='%s $\leq$ %.3f' % (label, l), node_size = 120)
    nx.draw_networkx_nodes(G0, pos=pos, nodelist=nodes_mid, node_color='#d80000',
                           label='%s $>$ %.3f, $\leq$ %.3f' % (label, l, h), node_size = 120)
    nx.draw_networkx_nodes(G0, pos=pos, nodelist=nodes_high, node_color='#4e0000',
                           label='%s $>$ %.3f' % (label, h), node_size = 120)
    nx.draw_networkx_edges(G0, pos=pos, width = 1)
    plt.legend(scatterpoints = 1)
    plt.title("Groups as Nodes")
    
    if appendix:
        plt.savefig('images/ch-appendicies/members/networkx_3cls_%s.png' % colname, bbox_inches = 'tight', pad_inches = 0.05)
        plt.close()
    else:
        plt.savefig('images/ch-members/active/networkx_3cls_%s.png' % colname, bbox_inches = 'tight', pad_inches = 0.05)
        plt.close()
    
    dists_low = [average_dists[i] for i in nodes_low]
    dists_mid = [average_dists[i] for i in nodes_mid]
    dists_high = [average_dists[i] for i in nodes_high]
    print("Three class graph")
    print('%s < %s Centrality: %.3f' %
          (label, l, np.mean(dists_low)))
    print('%s > %s, <= %s Centrality: %.3f' %
          (label, l, h, np.mean(dists_mid)))    
    print('%s > %s Centrality: %.3f' %
          (label, h, np.mean(dists_high)))
    print(f_oneway(dists_low, dists_mid, dists_high))
    print(ttest_ind(dists_low, dists_high))
    

threeClassGraphG0('Size', 'Size')

threeClassGraphG0('pCO', 'Prop. CO')

threeClassGraphG0('pVZ', 'Prop. VZ', True)

threeClassGraphG0('p3rdCountry', 'Prop. Non-CO, Non-VZ', True)

threeClassGraphG0('entropy', 'Entropy', True)

# Graph: Users as nodes

users_cols = [c for c in df_interact.columns if c[0] == '+']

matr_interact = df_interact[users_cols].fillna(0).values

G = nx.from_numpy_matrix(matr_interact)

Gcc = sorted(nx.connected_components(G), key=len, reverse=True)
G0 = G.subgraph(Gcc[0])

print("Users as nodes graph")
print("Largest CC n_nodes: %s" % G0.number_of_nodes())
print("Graph n_nodes: %s" % G.number_of_nodes())

def colorMap(tel):
    if tel[:4] == '+57 ':
        return '#8b0000'
    elif tel[:4] == '+58 ':
        return '#0000ff'
    else:
        return '#ffff00'

cmap = [colorMap(t) for t in df_interact.index.values]
cmap0 = [cmap[i] for i in list(Gcc[0])]

pos = nx.spring_layout(G0, iterations = 50, seed = 0)

nodes_co = [i for i in Gcc[0] if df_interact.index[i][:4] == '+57 ']
nodes_vz = [i for i in Gcc[0] if df_interact.index[i][:4] == '+58 ']
nodes_3rd = [i for i in Gcc[0] if (df_interact.index[i][:4] != '+57 ') and (df_interact.index[i][:4] != '+58 ')]

plt.figure(figsize = (14, 14))

nx.draw_networkx_nodes(G0, pos=pos, nodelist = nodes_co, node_size = 50, node_color = 'blue', alpha = 0.05)
nx.draw_networkx_nodes(G0, pos=pos, nodelist = nodes_vz, node_size = 50, node_color = 'red', alpha = 0.05)
nx.draw_networkx_nodes(G0, pos=pos, nodelist = [nodes_co[0]], node_size = 50, node_color = 'blue', alpha = 0.5, label = 'CO Users')
nx.draw_networkx_nodes(G0, pos=pos, nodelist = [nodes_vz[0]], node_size = 50, node_color = 'red', alpha = 0.5, label = 'VZ Users')

nx.draw_networkx_edges(G0, pos=pos, width = 0.005)
plt.legend()
plt.title("Users as Nodes (connected if in same group)")
plt.savefig('images/ch-members/active/networkx_co_vz.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.figure(figsize = (14, 14))

nx.draw_networkx_nodes(G0, pos=pos, nodelist = nodes_co, node_size = 50, node_color = 'blue', alpha = 0.05)
nx.draw_networkx_nodes(G0, pos=pos, nodelist = nodes_vz, node_size = 50, node_color = 'red', alpha = 0.05)
nx.draw_networkx_nodes(G0, pos=pos, nodelist = nodes_3rd, node_size = 50, node_color = 'yellow', alpha = 0.05)
nx.draw_networkx_nodes(G0, pos=pos, nodelist = [nodes_co[0]], node_size = 50, node_color = 'blue', alpha = 0.5, label = 'CO Users')
nx.draw_networkx_nodes(G0, pos=pos, nodelist = [nodes_vz[0]], node_size = 50, node_color = 'red', alpha = 0.5, label = 'VZ Users')
nx.draw_networkx_nodes(G0, pos=pos, nodelist = [nodes_3rd[0]], node_size = 50, node_color = 'yellow', alpha = 0.5, label = 'Other Users')

nx.draw_networkx_edges(G0, pos=pos, width = 0.005)
plt.legend()
plt.title("Users as Nodes (connected if in same group)")
plt.savefig('images/ch-members/active/networkx_co_vz_other.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

## Degree distributions of users

plt.figure(figsize = (12, 6))
df_interact['nConnections'].hist(bins = 100)
plt.xlabel("Number of Connections")
plt.ylabel('# of Users')
plt.title("User Degree Distribution (All Users)")
plt.savefig("images/ch-members/active/degree_dist.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.figure(figsize = (12, 6))
df_interact_co['nConnections'].hist(bins = 100, alpha = 0.3, range = (0, 500), label = 'CO Users')
df_interact_vz['nConnections'].hist(bins = 100, alpha = 0.4, range = (0, 500), label = 'VZ Users')
plt.legend()
plt.xlabel("Number of Connections")
plt.ylabel('# of Users')
plt.title("User Degree Distribution (CO and VZ Users)")
plt.savefig("images/ch-members/active/degree_dist_VZ_vs_CO.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print("nConnections means: VZ, CO")

print(df_interact_vz['nConnections'].mean())
print(df_interact_co['nConnections'].mean())

print(ttest_ind(df_interact_vz['nConnections'], df_interact_co['nConnections']))

print("Users with most connections")

print(df_interact['nConnections'].sort_values(ascending = False)[:20])

print("nConnections mean, median, prop > 100, prop > 200")

print(np.mean(df_interact['nConnections']))
print(np.median(df_interact['nConnections']))
print(np.mean(df_interact['nConnections'] > 100))
print(np.mean(df_interact['nConnections'] > 200))

df_groups.to_csv("Data/df_groups_nb1-members-byMessaging.csv")
