==> analysis/2_messages_overview_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta

import matplotlib.pyplot as plt
SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

import statsmodels.formula.api as smf
import scipy.stats
from scipy.stats import ttest_ind
from scipy.stats import f_oneway

df_x = pd.read_csv("Data/df_x_nb0a-groupMerge.csv", index_col = 0)
print(df_x.shape)

df_x['message_dt'] = pd.to_datetime(df_x["message_dt"], format='%Y-%m-%d %H:%M:00')
df_x['message_date'] = pd.to_datetime(df_x["message_date"], format='%Y-%m-%d')

df_x['text'] = df_x['text'].fillna('')
df_x['textlower'] = df_x['textlower'].fillna('')
df_x['image'] = df_x['image'].fillna('')
df_x['video_thumb'] = df_x['video_thumb'].fillna('')
df_x['emojis'] = df_x['emojis'].fillna('')
df_x['video_length'] = df_x['video_length'].fillna('')
df_x['audio_length'] = df_x['audio_length'].fillna('')

### Text similarity imports/functions

import string, re, unidecode
from wordcloud import WordCloud
from sklearn.decomposition import LatentDirichletAllocation as LDA

import nltk
from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer

from sklearn.feature_extraction.text import CountVectorizer

stemmer = SnowballStemmer('spanish')
remove_punc = str.maketrans(string.punctuation, len(string.punctuation) * " ")
stopwords_ascii = [unidecode.unidecode(w) for w in stopwords.words('spanish')]

def tokenize(s):
    s = unidecode.unidecode(s)
    s = s.translate(remove_punc)
    tokens = nltk.word_tokenize(s)
    filtered = [w for w in tokens if w not in stopwords_ascii]
    return [stemmer.stem(w) for w in filtered]

def dummy(x):
    return x

vectorizer = CountVectorizer(analyzer='word', tokenizer=dummy, preprocessor=dummy,
                            token_pattern=None)

# Descriptive statistics

print("Text: %s" % df_x[df_x['text'] != ''].shape[0])
print((df_x['text'] != '').mean())

print("Image: %s" % df_x[df_x['image'] != ''].shape[0])
print((df_x['image'] != '').mean())

print("Audio : %s" % df_x[df_x['audio_length'] != ''].shape[0])
print((df_x['audio_length'] != '').mean())

print("Video: %s" % df_x[df_x['video_thumb'] != ''].shape[0])
print((df_x['video_thumb'] != '').mean())

print("Emojis: %s" % df_x[df_x['emojis'] != ''].shape[0])
print((df_x['emojis'] != '').mean())

print(df_x.shape[0])

# Text

### Text length

word_counts = df_x[df_x['textlower'] != '']['textlower'].str.count(' ') + 1

df_x.loc[df_x['textlower'] != '', 'word_count'] = df_x[df_x['textlower'] != '']['textlower'].str.count(' ') + 1

plt.figure(figsize = (9, 5))
word_counts.hist(range = (0, 100), bins = 100)
plt.xlabel("Word Count")
plt.ylabel("# Messages")
plt.title("Text Messages")
plt.savefig('images/ch-messages/word_count.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(np.median(word_counts))
print(np.quantile(word_counts, .25))
print(np.quantile(word_counts, .75))
print(np.mean(word_counts == 1))
print(np.mean(word_counts < 20))
print(np.mean(word_counts > 100))
print(np.mean(word_counts > 500))

#### Group/user word count

group_word_count = df_x[df_x['word_count'] > 0].groupby('uid')['word_count'].mean()
user_word_count = df_x[df_x['word_count'] > 0].groupby('tel')['word_count'].mean()

group_word_count.hist(bins = 100)
plt.xlabel("Avg. Word Count")
plt.ylabel("# of Groups")
plt.savefig('images/ch-messages/group_word_count.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

user_word_count.hist(range = (100, 400), bins = 30)
plt.xlabel("Avg. Word Count")
plt.ylabel("# of Users")
plt.savefig('images/ch-messages/user_word_count.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

#### Char counts

char_counts = df_x[df_x['textlower'] != '']['text'].str.len()

plt.figure(figsize = (9, 5))
char_counts.hist(range = (0, 200), bins = 100)
plt.xlabel("Character Count")
plt.ylabel("# Messages")
plt.title("Text Messages")
plt.savefig('images/ch-messages/char_count.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(np.mean(char_counts))
print(np.median(char_counts))
print(np.quantile(char_counts, .25))
print(np.quantile(char_counts, .75))
print(np.mean(char_counts <= 10))
print(np.mean(char_counts > 522))

### LDA / WordCloud

stemmer = SnowballStemmer('spanish')
remove_punc = str.maketrans('', '', string.punctuation)
stopwords_ascii = [unidecode.unidecode(w) for w in stopwords.words('spanish')]

def tokenize(s):
    s = unidecode.unidecode(s)
    s = s.translate(remove_punc)
    tokens = nltk.word_tokenize(s)
    filtered = [w for w in tokens if w not in stopwords_ascii]
    return [stemmer.stem(w) for w in filtered]

def dummy(x):
    return x

count_vectorizer = CountVectorizer(analyzer='word', tokenizer=dummy, preprocessor=dummy,
                            token_pattern=None)

df_t = df_x[df_x['textlower'].str.len() > 0]
print(df_t.shape)

df_t.loc[:,'token'] = df_t.loc[:,'textlower'].apply(tokenize)

df_t['token_string'] = df_t['token'].apply(lambda x: ','.join(x))
all_texts = ','.join(list(df_t['token_string'].values))

wordcloud = WordCloud(background_color="white", max_words=5000, contour_width=3,
                      contour_color='steelblue', width=1400, height=800).generate(all_texts)
plt.figure(figsize = (14, 8))
plt.imshow(wordcloud, interpolation="bilinear")
plt.axis("off")
plt.savefig('images/ch-messages/wordcloud.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

X = count_vectorizer.fit_transform(df_t['token'])

# Tweak the two parameters below
number_topics = 10
number_words = 10

# Create and fit the LDA model
lda = LDA(n_components=number_topics, n_jobs=-1)
lda.fit(X)

# Helper function (sourced from somewhere online)
def print_topics(model, count_vectorizer, n_top_words):
    words = count_vectorizer.get_feature_names()
    for topic_idx, topic in enumerate(model.components_):
        print("\nTopic %d:" % topic_idx)
        print(" ".join([words[i]
                        for i in topic.argsort()[:-n_top_words - 1:-1]]))
        
# Print the topics found by the LDA model
print_topics(lda, count_vectorizer, number_words)

# Audio/video

def get_sec(t):
    if t.count(':') == 1:
        m, s = t.split(':')
        h = 0
    else:
        h, m, s = t.split(':')
    return int(h) * 3600 + int(m) * 60 + int(s)

audio_lengths_sec = df_x[df_x['audio_length'] != '']['audio_length'].apply(get_sec)

audio_lengths_sec.hist(bins = 100, range = (0, 400))
plt.xlabel("Length (seconds)")
plt.ylabel("# Messages")
plt.title("Audio Messages")
plt.savefig('images/ch-messages/audio_length.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(np.mean(audio_lengths_sec))
print(np.median(audio_lengths_sec))
print(np.quantile(audio_lengths_sec, .25))
print(np.quantile(audio_lengths_sec, .75))
print()
print(np.mean(audio_lengths_sec >= 30))
print(np.mean(audio_lengths_sec >= 60))
print(np.mean(audio_lengths_sec >= 300))
print()
print(np.mean((audio_lengths_sec >= 100) & (audio_lengths_sec < 200)))
print(np.mean((audio_lengths_sec >= 200) & (audio_lengths_sec < 300)))

video_lengths_sec = df_x[df_x['video_length'] != '']['video_length'].apply(get_sec)
video_lengths_sec_orig = \
    df_x[(df_x['video_length'] != '') & ~df_x['forwarded'] & ~df_x['forwarded_highly']]['video_length'].apply(get_sec)

video_lengths_sec.hist(bins = 100, range = (0, 400))
plt.xlabel("Length (seconds)")
plt.ylabel("# Messages")
plt.title("Video Messages")
plt.savefig('images/ch-messages/video_length.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(np.mean(video_lengths_sec))
print(np.median(video_lengths_sec))
print(np.quantile(video_lengths_sec, .25))
print(np.quantile(video_lengths_sec, .75))
print(np.mean(video_lengths_sec <= 10))
print(np.mean(video_lengths_sec >= 60))
print(np.mean(video_lengths_sec >= 300))
print()
print(np.mean(video_lengths_sec == 29))
print(np.mean(video_lengths_sec == 30))
print(np.mean(video_lengths_sec == 31))


print(np.mean(video_lengths_sec_orig == 29))
print(np.mean(video_lengths_sec_orig == 30))
print(np.mean(video_lengths_sec_orig == 31))

(df_x['forwarded'] | df_x['forwarded_highly']).mean()

print(((df_x['video_length'] != '') & (df_x['forwarded'] | df_x['forwarded_highly'])).sum())
print((df_x['video_length'] != '').sum())

print(((df_x['audio_length'] != '') & (df_x['forwarded'] | df_x['forwarded_highly'])).sum())
print((df_x['audio_length'] != '').sum())

# Activity/day

df_groups = pd.read_csv('Data/df_groups_nb1-members-byMessaging.csv', index_col = 0)

df_groups['n_days'] = \
    df_x[['uid', 'message_date']].groupby('uid').last() \
    - df_x[['uid', 'message_date']].groupby('uid').first() + timedelta(days = 1)
df_groups['n_days'] = df_groups['n_days'].dt.days

df_groups['activity'] = df_x.groupby('uid').size() / df_groups['n_days']

df_groups['activity'].hist(bins = 40)
plt.xlabel("Activity (Messages / Day)")
plt.ylabel("# of Groups")
plt.savefig('images/ch-messages/group_activity.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

df_groups[df_groups['activity'] < 20]['n_days'].hist(bins = 20, alpha = 0.5, label = 'Activity $< 20$')
df_groups[df_groups['activity'] >= 20]['n_days'].hist(bins = 20, alpha = 0.5, label = 'Activity $\geq 20$')
plt.xlabel("# of Days in Group")
plt.ylabel("# of Groups")
plt.title("Is being kicked out of a group\nendogenous to its activity?")
plt.legend()
plt.savefig('images/ch-messages/group_kicked.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

ttest_ind(df_groups[df_groups['activity'] < 20]['n_days'],
             df_groups[df_groups['activity'] >= 20]['n_days'])

### Correlates of activity

for col in ['Size', 'pVZ', 'pCO', 'pUS', 'pPE', \
       'pCL', 'pEC', 'p3rdCountry', 'entropy']:
    
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[col], df_groups['activity']))

reg = smf.ols('activity ~ Size + entropy', data = df_groups).fit()
print(reg.summary())

plt.scatter(df_groups['Size'], df_groups['activity'], alpha = 0.3)
plt.xlabel("Group Size")
plt.ylabel("Group Activity")
plt.title("All Groups")
plt.savefig('images/ch-messages/scatter_size_activity.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.scatter(df_groups['entropy'], df_groups['activity'], alpha = 0.3)
plt.xlabel("Group Entropy")
plt.ylabel("Group Activity")
plt.title("All Groups")
plt.savefig('images/ch-messages/scatter_entropy_activity.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

df_groups.to_csv('Data/df_groups_nb2-messages.csv')
