==> analysis/2a_messages_concentration_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta

import matplotlib.pyplot as plt
SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

import statsmodels.formula.api as smf
import scipy.stats

df_x = pd.read_csv("Data/df_x_nb0a-groupMerge.csv", index_col = 0)
print(df_x.shape)

# Concentration

def hhConc(df):
    counts = df['tel'].value_counts()
    props = counts / sum(counts)
    
    return sum(props**2)

def top5Conc(df):
    counts = df['tel'].value_counts()
    props = counts / sum(counts)
    
    return sum(props[:5])

def gini(df):
    counts = np.sort(df['tel'].value_counts())
    n = len(counts)
    index = np.arange(1, n + 1)
    
    return sum((2 * index - n  - 1) * counts) / (n * sum(counts))

df_groups = pd.read_csv('Data/df_groups_nb2-messages.csv', index_col = 0)

df_groups['hhConc'] = df_x[['uid', 'tel']].groupby('uid').apply(hhConc)
df_groups['gini'] = df_x[['uid', 'tel']].groupby('uid').apply(gini)

df_groups = df_groups.drop(['simpson', 'G0_avg_dist', 'clustering', 'n_days'], axis = 1)

df_groups.to_csv('Data/df_groups_nb2b-concentration.csv')

df_groups['top5Conc'] = df_x[['uid', 'tel']].groupby('uid').apply(top5Conc)

plt.scatter(df_groups['hhConc'], df_groups['top5Conc'], alpha = 0.3)
plt.xlabel("H-H Concentration")
plt.ylabel("Top 5 Concentration")
plt.title("All Groups")
plt.savefig('images/ch-messages/scatter_hh_top5.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(scipy.stats.pearsonr(df_groups['hhConc'], df_groups['top5Conc']))

plt.scatter(df_groups['hhConc'], df_groups['gini'], alpha = 0.3)
plt.xlabel("H-H Concentration")
plt.ylabel("Gini Coefficient")
plt.title("All Groups")
m, b = np.polyfit(df_groups['hhConc'], df_groups['gini'], 1)
plt.plot(df_groups['hhConc'], m*df_groups['hhConc'] + b, color = 'orange')
plt.savefig('images/ch-messages/scatter_hh_gini.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(scipy.stats.pearsonr(df_groups['hhConc'], df_groups['gini']))

plt.scatter(df_groups[df_groups['Size'] > 1]['hhConc'], df_groups[df_groups['Size'] > 1]['gini'], alpha = 0.3)
plt.xlabel("H-H Concentration")
plt.ylabel("Gini Coefficient")
plt.title("Groups With $> 1$ Member")
m, b = np.polyfit(df_groups[df_groups['Size'] > 1]['hhConc'], df_groups[df_groups['Size'] > 1]['gini'], 1)
plt.plot(df_groups[df_groups['Size'] > 1]['hhConc'], m*df_groups[df_groups['Size'] > 1]['hhConc'] + b, color = 'orange')
plt.savefig('images/ch-messages/scatter_hh_gini_drop.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(scipy.stats.pearsonr(df_groups[df_groups['Size'] > 1]['hhConc'], df_groups[df_groups['Size'] > 1]['gini']))

print((df_groups['Size'] == 1).sum())

## hhConc correlations

df_groups.columns

for col in ['Size', 'pVZ', 'pCO', 'pUS', 'pPE', \
       'pCL', 'pEC', 'p3rdCountry', 'entropy', 'degree', 'activity']:
    
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[col], df_groups['hhConc']))

reg = smf.ols('hhConc ~ Size + entropy + degree + activity', data = df_groups).fit()
print(reg.summary())

plt.scatter(df_groups['Size'], df_groups['hhConc'])
plt.xlabel("Size")
plt.ylabel("H-H Concentration")
m, b = np.polyfit(df_groups['Size'], df_groups['hhConc'], 1)
plt.plot(range(0, 200), m*range(0, 200) + b, color = 'orange')
plt.savefig("images/ch-messages/scatter_size_hh.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.scatter(df_groups['entropy'], df_groups['hhConc'])
plt.xlabel("Entropy")
plt.ylabel("H-H Concentration")
m, b = np.polyfit(df_groups['entropy'], df_groups['hhConc'], 1)
plt.plot(df_groups['entropy'], m*df_groups['entropy'] + b, color = 'orange')
plt.savefig("images/ch-messages/scatter_entropy_hh.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

## Gini correlations

for col in ['Size', 'pVZ', 'pCO', 'pUS', 'pPE', \
       'pCL', 'pEC', 'p3rdCountry', 'entropy', 'activity', 'degree']:
    
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[col], df_groups['gini']))

reg = smf.ols('gini ~ Size + entropy + activity + degree', data = df_groups).fit()
print(reg.summary())

plt.scatter(df_groups['Size'], df_groups['gini'])
plt.xlabel("Size")
plt.ylabel("Gini")
m, b = np.polyfit(df_groups['Size'], df_groups['gini'], 1)
plt.plot(range(0, 300), m*range(0, 300) + b, color = 'orange')
plt.savefig("images/ch-messages/scatter_size_gini.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.scatter(df_groups['entropy'], df_groups['gini'])
plt.xlabel("Entropy")
plt.ylabel("Gini")
m, b = np.polyfit(df_groups['entropy'], df_groups['gini'], 1)
plt.plot(df_groups['entropy'], m*df_groups['entropy'] + b, color = 'orange')
plt.savefig("images/ch-messages/scatter_entropy_gini.png", bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
