==> analysis/2b_messages_repeated_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta

import matplotlib.pyplot as plt
SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

import statsmodels.formula.api as smf
from scipy.stats import ttest_ind
from collections import Counter

df_x = pd.read_csv("Data/df_x_nb0a-groupMerge.csv", index_col = 0)
print(df_x.shape)

df_x['textlower'] = df_x['textlower'].fillna('')
df_x['image'] = df_x['image'].fillna('')
df_x['video_thumb'] = df_x['video_thumb'].fillna('')
df_x['video_length'] = df_x['video_length'].fillna('')
df_x['audio_length'] = df_x['audio_length'].fillna('')

df_x['message_dt'] = pd.to_datetime(df_x["message_dt"], format='%Y-%m-%d %H:%M:00')

df_groups = pd.read_csv('Data/df_groups_nb2b-concentration.csv', index_col = 0)

# Repeated images

df_x['image_list'] = df_x['image'].apply(lambda x: [i for i in x.split('|') if i != ''])
all_images = df_x['image_list'].sum()

image_counter = Counter(all_images)

image_shares = pd.DataFrame.from_dict(image_counter, orient='index', columns=['n_shares'])

print(len(image_shares))

### n_shares

image_shares[image_shares['n_shares'] >= 2]['n_shares'].hist(bins = 40)
plt.xlabel("Number of Shares")
plt.ylabel("# of Images")
plt.title("Images Shared Twice or More")
plt.savefig('images/ch-messages/hist_image_nshare.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(np.mean(image_shares['n_shares'] == 1))
print(np.sum(image_shares['n_shares'] == 1))
print(np.sum(image_shares['n_shares'] > 1))
print()
print(np.sum(image_shares['n_shares'] == 2))
print(np.mean(image_shares['n_shares'] <= 5))

### Images shared over 5 times

groups_anyshare = set()
groups_firstshare = set()

for image in image_shares[image_shares['n_shares'] > 5].index:
    groups_anyshare.update(df_x[df_x['image'].str.contains(image)]['uid'].unique())
    groups_firstshare.add(df_x[df_x['image'].str.contains(image)]['uid'].iloc[0])

print(image_shares[image_shares['n_shares'] > 5].shape[0])
print(len(groups_anyshare))
print(len(groups_firstshare))
print(df_groups.shape[0])

df_groups.columns

stat_cols = ['Size', 'pVZ', 'pCO', 'pUS', 'pPE',
       'pCL', 'pEC', 'p3rdCountry', 'entropy', 'degree', 'activity', 'hhConc', 'gini']

df_groups_anyshare = df_groups.loc[df_groups.index.isin(groups_anyshare)][stat_cols]
df_groups_firstshare = df_groups.loc[df_groups.index.isin(groups_firstshare)][stat_cols]

df_groups_notany = df_groups.loc[~df_groups.index.isin(groups_anyshare)][stat_cols]
df_groups_notfirst = df_groups.loc[~df_groups.index.isin(groups_firstshare)][stat_cols]

for col in stat_cols:
    print(col)
    print(np.mean(df_groups_firstshare[col]), np.mean(df_groups_notfirst[col]))
    print(ttest_ind(df_groups_firstshare[col], df_groups_notfirst[col]))
    print()

for col in stat_cols:
    print(col)
    print(np.mean(df_groups_anyshare[col]), np.mean(df_groups_notany[col]))
    print(ttest_ind(df_groups_anyshare[col], df_groups_notany[col]))
    print()

### Regressing number of shares on group characteristics

def firstGroupAppearImage(image):
    return df_x.loc[df_x.image.str.contains(image), 'uid'].iloc[0]

image_shares['first_group'] = image_shares.index.map(firstGroupAppearImage).values

df_groups_Size = df_groups['Size'].to_dict()
df_groups_entropy = df_groups['entropy'].to_dict()
df_groups_activity = df_groups['activity'].to_dict()
df_groups_hhConc = df_groups['hhConc'].to_dict()
df_groups_gini = df_groups['gini'].to_dict()
df_groups_degree = df_groups['degree'].to_dict()

image_shares['Size'] = image_shares['first_group'].apply(lambda x: df_groups_Size[x])
image_shares['entropy'] = image_shares['first_group'].apply(lambda x: df_groups_entropy[x])
image_shares['activity'] = image_shares['first_group'].apply(lambda x: df_groups_activity[x])
image_shares['hhConc'] = image_shares['first_group'].apply(lambda x: df_groups_hhConc[x])
image_shares['gini'] = image_shares['first_group'].apply(lambda x: df_groups_gini[x])
image_shares['degree'] = image_shares['first_group'].apply(lambda x: df_groups_degree[x])

image_shares.to_csv('Data/intermediate/df_image_shares_nb2a-repeated.csv')

#### Re-import data from last run

image_shares = pd.read_csv('Data/intermediate/df_image_shares_nb2a-repeated.csv', index_col = 0)

image_shares

reg = smf.ols('n_shares ~ Size + entropy + degree + activity + hhConc + gini',
              data = image_shares).fit()
print(reg.summary())

reg = smf.ols('n_shares ~ Size + entropy + degree + activity + hhConc + gini',
              data = image_shares[image_shares['n_shares'] > 1]).fit()
print(reg.summary())

### Image share time

def image_shareTimerange(image):
    times = df_x[df_x['image'].str.contains(image)]['message_dt']
    return (times.iloc[-1] - times.iloc[0]).total_seconds()

image_shares.loc[image_shares['n_shares'] > 1, 'timerange'] = image_shares.loc[image_shares['n_shares'] > 1].index.map(image_shareTimerange)
image_shares['timerange'] = image_shares['timerange'] / 3600
image_shares['timerange'] = image_shares['timerange'].fillna(0)

plt.figure()
image_shares[image_shares['timerange'] > 0]['timerange'].hist(bins = 50)
plt.xlabel("Share Time Range (Hours)")
plt.ylabel("# of Images")
plt.title("Images Shared Twice or More")
plt.savefig('images/ch-messages/hist_image_timerange.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

reg = smf.ols('timerange ~ Size + entropy + degree + activity + hhConc + gini',
              data = image_shares).fit()
print(reg.summary())

reg = smf.ols('timerange ~ Size + entropy + degree + activity + hhConc + gini',
              data = image_shares[image_shares['timerange'] > 0]).fit()
print(reg.summary())

# Repeated text

text_counter = Counter(df_x['textlower'])
text_shares = pd.DataFrame.from_dict(text_counter, orient='index', columns=['n_shares'])
text_shares = text_shares.drop('')
text_shares = text_shares.iloc[text_shares.index.str.len() >= 20]

print(np.mean(text_shares['n_shares'] == 1))
print(np.mean(text_shares['n_shares'] <= 2))
print(np.mean(text_shares['n_shares'] <= 5))

Not interesting!

# Video

video_counter = Counter(df_x['video_thumb'] + df_x['video_length'])
video_shares = pd.DataFrame.from_dict(video_counter, orient='index', columns=['n_shares'])
video_shares = video_shares.drop('')

print(np.mean(video_shares['n_shares'] == 1))
print(np.mean(video_shares['n_shares'] == 2))
print(np.mean(video_shares['n_shares'] <= 5))

print(len(video_shares))

### Identify first group and characteristics

def firstGroupAppearVideo(videoLength):
    return df_x.loc[(df_x['video_thumb'] + df_x['video_length'] == videoLength), 'uid'].iloc[0]

video_shares['first_group'] = video_shares.index.map(firstGroupAppearVideo).values

video_shares['Size'] = video_shares['first_group'].apply(lambda x: df_groups_Size[x])
video_shares['entropy'] = video_shares['first_group'].apply(lambda x: df_groups_entropy[x])
video_shares['activity'] = video_shares['first_group'].apply(lambda x: df_groups_activity[x])
video_shares['hhConc'] = video_shares['first_group'].apply(lambda x: df_groups_hhConc[x])
video_shares['gini'] = video_shares['first_group'].apply(lambda x: df_groups_gini[x])

video_shares.to_csv('Data/intermediate/df_video_shares_nb2a-repeated.csv')

#### Import from previous

video_shares = pd.read_csv('Data/intermediate/df_video_shares_nb2a-repeated.csv', index_col = 0)

video_shares

## Regression

reg = smf.ols('n_shares ~ Size + entropy + degree + activity + hhConc + gini',
              data = video_shares[video_shares['n_shares'] > 1]).fit()
print(reg.summary())

### Time range

def video_shareTimerange(videoLength):
    times = df_x.loc[(df_x['video_thumb'] + df_x['video_length'] == videoLength), 'message_dt']
    return (times.iloc[-1] - times.iloc[0]).total_seconds()

video_shares.loc[video_shares['n_shares'] > 1, 'timerange'] = video_shares.loc[video_shares['n_shares'] > 1].index.map(video_shareTimerange)
video_shares['timerange'] = video_shares['timerange'] / 3600
video_shares['timerange'] = video_shares['timerange'].fillna(0)

reg = smf.ols('timerange ~ Size + entropy + degree + activity + hhConc + gini',
              data = video_shares[video_shares['n_shares'] > 1]).fit()
print(reg.summary())
