==> analysis/3a_replyCascades_overview_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta, time

import ast # string rep. of list to list

import statsmodels.formula.api as smf
import scipy.stats
from scipy.stats import ttest_ind

import matplotlib.pyplot as plt
import matplotlib.dates as mdates
SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

df_x = pd.read_csv('Data/df_x_nb3-reply.csv', index_col=0)
print(df_x.shape)

df_x['reply_link'] = df_x['reply_link'].astype(pd.Int32Dtype())
df_x['reply_list'] = df_x['reply_list'].apply(lambda x: ast.literal_eval(x))

df_x['message_dt'] = pd.to_datetime(df_x["message_dt"], format='%Y-%m-%d %H:%M:00')
df_x['message_date'] = pd.to_datetime(df_x["message_date"], format='%Y-%m-%d')

# General statistics

df_x.shape[0]

# Traced replies
trace = df_x[df_x['reply_link'] > 0].shape[0]
actual = df_x[df_x['reply_to'].notnull()].shape[0]
print(trace)
print(actual)
print(trace/actual)

## Messages that were replied to

# Replied to
df_x[df_x['replies_n'] > 0].shape[0]

df_x[df_x['replies_n'] > 0]['replies_n'].hist(range = (1, 10), bins = 10)
plt.xlabel("Number of Replies Received")
plt.ylabel("# of Messages")
plt.savefig('images/ch-replycascades/hist_replies_n.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

df_x[df_x['replies_n'] == 1].shape[0]

df_x[df_x['replies_n'] == 2].shape[0]

df_x[(df_x['replies_n'] > 0) & (df_x['replies_n'] < 5)].shape[0]

### Content type

withrep = df_x[(df_x['replies_n'] > 0) & (df_x['text'].notnull())].shape[0]
n = df_x[df_x['text'].notnull()].shape[0]

print(withrep)
print(n)
print(withrep/n)

withrep = df_x[(df_x['replies_n'] > 0) & (df_x['image'].notnull())].shape[0]
n = df_x[df_x['image'].notnull()].shape[0]

print(withrep)
print(n)
print(withrep/n)

withrep = df_x[(df_x['replies_n'] > 0) & (df_x['video_length'].notnull())].shape[0]
n = df_x[df_x['video_length'].notnull()].shape[0]

print(withrep)
print(n)
print(withrep/n)

withrep = df_x[(df_x['replies_n'] > 0) & (df_x['emojis'].notnull())].shape[0]
n = df_x[df_x['emojis'].notnull()].shape[0]

print(withrep)
print(n)
print(withrep/n)

withrep = df_x[(df_x['replies_n'] > 0) & (df_x['audio_length'].notnull())].shape[0]
n = df_x[df_x['audio_length'].notnull()].shape[0]

print(withrep)
print(n)
print(withrep/n)

### Forwarded content

withrep = df_x[(df_x['replies_n'] > 0) & (df_x['forwarded'] | df_x['forwarded_highly'])].shape[0]
n = df_x[(df_x['forwarded'] | df_x['forwarded_highly'])].shape[0]

print(withrep)
print(n)
print(withrep/n)

# not forwarded images
# SLIGHTLY HIGHER
withrep = df_x[(df_x['replies_n'] > 0) & ~(df_x['forwarded'] | df_x['forwarded_highly']) \
    & (df_x['image'].notnull())].shape[0]
n = df_x[~(df_x['forwarded'] | df_x['forwarded_highly']) \
    & (df_x['image'].notnull())].shape[0]

print(withrep)
print(n)
print(withrep/n)

## By nationality

df_c = df_x[df_x.tel.str.startswith('+57')]

df_v = df_x[df_x.tel.str.startswith('+58')]

for col in ['image', 'text', 'video_length', 'audio_length']:
    print(col)
    print(df_c[df_c[col].notnull()].shape[0] / df_c.shape[0])
    print(df_v[df_v[col].notnull()].shape[0] / df_v.shape[0])
    print()

print((df_c['replies_n'] > 0).mean())
print((df_v['replies_n'] > 0).mean())

print(ttest_ind(df_c['replies_n'] > 0, df_v['replies_n'] > 0))
print(ttest_ind(df_c['image'].notnull(), df_v['image'].notnull()))
print(ttest_ind(df_c['text'].notnull(), df_v['text'].notnull()))

# By time of day

df_x['round_hour'] = df_x['message_dt'].dt.round('30T').apply(lambda x: x.time())

x = [(pd.Timestamp(2017, 1, 1, h), pd.Timestamp(2017, 1, 1, h, 30)) for h in range(0, 24)]
x = [j for i in x for j in i]
y = df_x[['round_hour', 'replies_n']].groupby('round_hour').mean().values
y_text = df_x[df_x['text'].notnull()][['round_hour', 'replies_n']].groupby('round_hour').mean().values

fig, ax = plt.subplots(figsize = (6, 4))
ax.plot(x, y, label = 'All Messages')
ax.plot(x, y_text, label = 'Text Messages')

ax.xaxis.set_major_locator(mdates.HourLocator(interval = 8))
ax.xaxis.set_major_formatter(mdates.DateFormatter('%H:%M:%S'))

fig.autofmt_xdate()
plt.xticks(ha='center', rotation=0)
plt.xlabel("Time Message Is Sent\n(30 Minute Buckets)")
plt.ylabel("Avg. # of\nReplies Received")
plt.legend()
plt.savefig('images/ch-replycascades/time_of_day.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

def replyToNotNull(df):
    return df['reply_to'].notnull().mean()

y = df_x[['round_hour', 'reply_to']].groupby('round_hour').apply(replyToNotNull).values

fig, ax = plt.subplots(figsize = (6, 4))
ax.plot(x, y * 100)
ax.xaxis.set_major_locator(mdates.HourLocator(interval = 8))
ax.xaxis.set_major_formatter(mdates.DateFormatter('%H:%M:%S'))

fig.autofmt_xdate()
plt.xticks(ha='center', rotation=0)
plt.ylabel("What % of messages\nare replies?")
plt.savefig('images/ch-replycascades/time_of_day_areReplies.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# Average replies_n within groups

df_groups = pd.read_csv('Data/df_groups_nb2b-concentration.csv', index_col = 0)

df_groups['replies_n'] = df_x[['uid', 'replies_n']].groupby('uid').mean()

df_groups.to_csv('Data/df_groups_nb3a-reply.csv')

print((df_groups['replies_n'] == 0).sum())

df_groups['replies_n'].hist(bins = 20)
plt.xlabel("Avg. Number of Replies\n(To All Messages)")
plt.ylabel("# of Groups")
plt.savefig('images/ch-replycascades/hist_group_replies_n.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

reg = smf.ols('replies_n ~ entropy + hhConc + gini', data = df_groups).fit()
print(reg.summary())
# strong in HHCONC

reg = smf.ols('replies_n ~ entropy + hhConc + gini',
              data = df_groups[df_groups['replies_n'] > 0]).fit()
print(reg.summary())
# strong in HHCONC

cols = ['hhConc', 'entropy']
labs = ['H-H Concentration', 'Entropy']

for c, l in zip(cols, labs):
    plt.scatter(df_groups[c], df_groups['replies_n'], alpha = 0.3)
    plt.xlabel(l)
    plt.ylabel("Average # of Replies")
    plt.title("All Groups")
    m, b = np.polyfit(df_groups[c], df_groups['replies_n'], 1)
    plt.plot(df_groups[c], m*df_groups[c] + b, color = 'orange')
    plt.savefig('images/ch-replycascades/scatter_%s_replies_n.png' % c, bbox_inches = 'tight', pad_inches = 0.05)
    plt.close()

## Competing messages

def n_msg_10msurround(row):
    return df_x[(df_x.uid == row.uid) \
     & (df_x.message_dt <= row.message_dt + timedelta(minutes = 5)) \
     & (df_x.message_dt >= row.message_dt - timedelta(minutes = 5))].shape[0] - 1

df_x.loc[df_x['replies_n'] > 0, 'n_msg_10msurround'] = df_x.apply(n_msg_10msurround, axis = 1)

df_x.to_csv('Data/intermediate/df_x_nb3a-surround.csv')

### Reimport

df_x = pd.read_csv('Data/intermediate/df_x_nb3a-surround.csv')

df_x['n_msg_10msurround'].hist(bins = 20)

plt.title("Messages with Replies")
plt.xlabel('Number of "Competing" Messages\nin $[t-5, t+5]$ Window')
plt.ylabel('# of Messages')
plt.savefig('images/ch-replycascades/competing_msg.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

(df_x.loc[df_x['replies_n'] > 0, 'n_msg_10msurround'] < 10).mean()

(df_x.loc[df_x['replies_n'] > 0, 'n_msg_10msurround'] < 20).mean()
