==> analysis/3b_replyCascades_virality-theory_PRODUCTION.ipynb <==
import numpy as np
import matplotlib.pyplot as plt

SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

def genBinaryTree(n):
    graph = {}
    
    for i in range(int(n/2)):
        graph[i] = graph.get(i, []) + [2*i + 1, 2*i + 2]
        graph[2*i + 1] = [i]
        graph[2*i + 2] = [i]
    
    return graph

def avgDistBinary_GOEL(n):
    g = genBinaryTree(n)

    dist_matrix = np.zeros((n,n))

    for i in range(n-1):
        # BFS
        discovered = [False] * n
        distances = [0] * n

        # Label start
        queue = [i]
        discovered[i] = True

        while len(queue) > 0:
            v = queue.pop(0)

            neighbors = g[v]
            for ni in neighbors:
                if not discovered[ni]:
                    discovered[ni] = True
                    queue.append(ni)
                    distances[ni] = distances[v] + 1
        dist_matrix[i] = distances
        dist_matrix[:, i] = distances

    return np.sum(dist_matrix) / (n * (n-1))

def broadcastDist_GOEL(n):
    return 2 * (n-1) / n

sizes = [2**k - 1 for k in range(2, 11)]
print(sizes)

plt.figure(figsize = (8, 5))
plt.plot(sizes, [avgDistBinary_GOEL(s) for s in sizes], label = 'Binary Tree', linewidth=4)
plt.plot(range(3, 1024), [broadcastDist_GOEL(s) for s in range(3, 1024)], label = 'Broadcast', linewidth=4)
plt.xlabel("Size of Cascade")
plt.ylabel("Avg. Dist. between Nodes")
plt.legend()
plt.savefig('images/ch-replycascades/binary_vs_broadcast.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

#### Our measure

def avgDistBinary_CHANG(n):
    g = genBinaryTree(n)

    dist_matrix = np.zeros((n,n))

    for i in range(n-1):
        # BFS
        discovered = [False] * n
        distances = [0] * n

        # Label start
        queue = [i]
        discovered[i] = True

        while len(queue) > 0:
            v = queue.pop(0)

            neighbors = g[v]
            for ni in neighbors:
                if not discovered[ni]:
                    discovered[ni] = True
                    queue.append(ni)
                    distances[ni] = distances[v] + 1
        dist_matrix[i] = distances
        dist_matrix[:, i] = distances

    return np.mean(dist_matrix)

def broadcastDist_CHANG(n):
    return 2 * (n-1)**2 / (n**2)

sizes2 = [1, 3, 7, 15]

plt.figure(figsize = (12, 6))
plt.plot(sizes2, [avgDistBinary_GOEL(s) for s in sizes2], label = 'Binary Tree (Goel)', alpha = 0.5, linestyle='dashed', linewidth=4)
plt.plot(range(1,16), [broadcastDist_GOEL(s) for s in range(1,16)], label = 'Broadcast (Goel)', alpha = 0.5, linestyle='dashed', linewidth=4)
plt.plot(sizes2, [avgDistBinary_CHANG(s) for s in sizes2], label = 'Binary Tree (Ours)', linewidth=4)
plt.plot(range(1,16), [broadcastDist_CHANG(s) for s in range(1,16)], label = 'Broadcast (Ours)', linewidth=4)
plt.xlabel("Size of Cascade")
plt.ylabel("Avg. Dist. between Nodes")
plt.legend()
plt.savefig('images/ch-replycascades/chang_vs_goel.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
