==> analysis/4_misinformation_extractManual_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd
import os
import ast

import hashlib

df_x = pd.read_csv("Data/df_x_nb3-reply.csv", index_col = 0)
print(df_x.shape)

# Flagging

df_x['hash'] = df_x['textlower'].apply(lambda x: hashlib.md5(x.encode('utf-8')).hexdigest())

vc = df_x['hash'].value_counts()
for hs in vc[vc >= 3].index.tolist():
    t = df_x[df_x['hash'] == hs].iloc[0]['textlower']
    if len(t) >= 20:
        print(t + '\n')
        print(hs + '\n--------------------')
