==> analysis/4b_misinformation_overview-fakeNews_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

import statsmodels.formula.api as smf
import scipy.stats
from scipy.stats import ttest_ind
from scipy.stats import f_oneway

df_x = pd.read_csv("Data/df_x_nb4a-mis.csv", index_col = 0)
print(df_x.shape)

df_x['text'] = df_x['text'].fillna('')
df_x['textlower'] = df_x['textlower'].fillna('')
df_x['virality'] = df_x['virality'].fillna(0)

### Text similarity imports/functions

import string, re, unidecode
from sklearn.decomposition import LatentDirichletAllocation as LDA

import nltk
from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer

from collections import Counter

from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity

stemmer = SnowballStemmer('spanish')
remove_punc = str.maketrans(string.punctuation, len(string.punctuation) * " ")
stopwords_ascii = [unidecode.unidecode(w) for w in stopwords.words('spanish')]

def tokenize(s):
    s = unidecode.unidecode(s)
    s = s.translate(remove_punc)
    tokens = nltk.word_tokenize(s)
    filtered = [w for w in tokens if w not in stopwords_ascii]
    return [stemmer.stem(w) for w in filtered]

def dummy(x):
    return x

vectorizer = TfidfVectorizer(analyzer='word', tokenizer=dummy, preprocessor=dummy,
                            token_pattern=None)

count_vectorizer = CountVectorizer(analyzer='word', tokenizer=dummy, preprocessor=dummy,
                            token_pattern=None)

df_groups = pd.read_csv('Data/df_groups_nb3b-virality.csv', index_col = 0)
print(df_groups.head())

# Initial statistics

# Message dynamics

df_x['posMisinfo'] = (df_x['textlower'].apply(lambda x: len(tokenize(x)) >= 5))
df_p = df_x[df_x['posMisinfo']]

print(df_p[df_p['fake_news']].shape)
print(df_p.shape)

### Replies

reg = smf.ols('replies_n ~ fake_news', data = df_p).fit()
print(reg.summary())
reg = smf.ols('replies_n ~ fake_news', data = df_x).fit()
print(reg.summary())

print(np.quantile(df_p[df_p['fake_news']]['replies_n'], .95))
print(np.quantile(df_p[~df_p['fake_news']]['replies_n'], .95))

print(np.mean(df_p[df_p['fake_news']]['replies_n'] > 0))

### Virality

reg = smf.ols('virality ~ fake_news', data = df_p[df_p['virality'] > 0]).fit()
print(reg.summary())
reg = smf.ols('virality ~ fake_news', data = df_x[df_x['virality'] > 0]).fit()
print(reg.summary())

print(np.quantile(df_p[df_p['fake_news'] & (df_p['virality'] > 0)]['virality'], .95))

print(np.quantile(df_p[~df_p['fake_news'] & (df_p['virality'] > 0)]['virality'], .95))

### Size

df_p['charlength'] = df_p['text'].apply(len)
df_p['wordlength'] = df_p['text'].apply(lambda x: x.count(" "))

reg = smf.ols('charlength ~ fake_news', data = df_p).fit()
print(reg.summary())
reg = smf.ols('wordlength ~ fake_news', data = df_p).fit()
print(reg.summary())

### LDA

df_t = df_p[df_p['fake_news']]
df_t.loc[:,'token'] = df_t.loc[:,'textlower'].apply(tokenize)

X = count_vectorizer.fit_transform(df_t['token'])

# Tweak the two parameters below
number_topics = 10
number_words = 10

# Create and fit the LDA model
lda = LDA(n_components=number_topics, n_jobs=-1)
lda.fit(X)

# Helper function (sourced from somewhere online)
def print_topics(model, count_vectorizer, n_top_words):
    words = count_vectorizer.get_feature_names()
    for topic_idx, topic in enumerate(model.components_):
        print("\nTopic %d:" % topic_idx)
        print(" ".join([words[i]
                        for i in topic.argsort()[:-n_top_words - 1:-1]]))
        
# Print the topics found by the LDA model
print_topics(lda, count_vectorizer, number_words)

# User dynamics

df_x[df_x['fake_news']]['tel'].value_counts().hist(bins = 20)
plt.xlabel("Number of Fake News Shared")
plt.ylabel("# of Users")
plt.savefig('images/ch-misinformation/hist_fakenews_user_frequency.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(df_x[df_x['fake_news']]['tel'].unique().shape)

print(np.mean(df_x[df_x['fake_news']]['tel'].value_counts() == 1))
print(np.mean(df_x[df_x['fake_news']]['tel'].value_counts() == 2))

tel_fake_news = df_p[['tel', 'fake_news']].groupby('tel').mean()
tel_fake_news_n = df_p[['tel', 'fake_news']].groupby('tel').sum()

print(np.mean(tel_fake_news == 0))
print(tel_fake_news.shape)

tel_fake_news[tel_fake_news > 0].hist(bins = 20)
plt.xlabel("Prevalence of Fake News")
plt.ylabel("# of Users")
plt.title(None)
plt.savefig('images/ch-misinformation/hist_fakenews_user_prevalence.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

tel_fake_news_n[tel_fake_news_n['fake_news'] >= 5]

## Comparing by country

### User shares fake news

x = ['PER', 'CHL', 'COL', 'VEN', 'ECU']

pe = tel_fake_news[tel_fake_news.index.str.startswith('+51')] > 0
cl = tel_fake_news[tel_fake_news.index.str.startswith('+56')] > 0
co = tel_fake_news[tel_fake_news.index.str.startswith('+57')] > 0
vz = tel_fake_news[tel_fake_news.index.str.startswith('+58')] > 0
ec = tel_fake_news[tel_fake_news.index.str.startswith('+593')] > 0

y = [100 * np.mean(country)[0] for country in [pe, cl, co, vz, ec]]

plt.figure()
plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("% of Users Who've\nShared Fake News")
plt.savefig('images/ch-misinformation/bar_fakenews_country_whoshared.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

print(np.mean(vz))
print(np.mean(co))

### Prop of user content is fake news

pe = tel_fake_news[tel_fake_news.index.str.startswith('+51')]
cl = tel_fake_news[tel_fake_news.index.str.startswith('+56')]
co = tel_fake_news[tel_fake_news.index.str.startswith('+57')]
vz = tel_fake_news[tel_fake_news.index.str.startswith('+58')]
ec = tel_fake_news[tel_fake_news.index.str.startswith('+593')]

y = [np.mean(country)[0] for country in [pe, cl, co, vz, ec]]

plt.figure()
plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("Avg. Fake News Prevalence")
plt.title("All Users from Country")
plt.savefig('images/ch-misinformation/bar_fakenews_country_prevalence.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

print(np.mean(vz))
print(np.mean(co))

### Prop user content is fake news, ONLY FOR FAKE NEWS sharers

tel_fake_news_pos = tel_fake_news[tel_fake_news['fake_news'] > 0]

pe = tel_fake_news_pos[tel_fake_news_pos.index.str.startswith('+51')]['fake_news'].values
cl = tel_fake_news_pos[tel_fake_news_pos.index.str.startswith('+56')]['fake_news'].values
co = tel_fake_news_pos[tel_fake_news_pos.index.str.startswith('+57')]['fake_news'].values
vz = tel_fake_news_pos[tel_fake_news_pos.index.str.startswith('+58')]['fake_news'].values
ec = tel_fake_news_pos[tel_fake_news_pos.index.str.startswith('+593')]['fake_news'].values

fig, ax = plt.subplots()
ax.violinplot([pe, cl, co, vz, ec], showmeans = True)
ax.set_xticklabels(['','PER', 'CHL', 'COL', 'VEN', 'ECU'])
plt.xlabel("Country")
plt.ylabel("Fake News Prevalence")
plt.title("Users Who've Shared Fake News")
plt.savefig('images/ch-misinformation/violin_fakenews_country_prevalence_sharers.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

### Number of fake news

print(np.mean(tel_fake_news_n[tel_fake_news_n.index.str.startswith('+57')]))
print(np.mean(tel_fake_news_n[tel_fake_news_n.index.str.startswith('+58')]))
ttest_ind(tel_fake_news_n[tel_fake_news_n.index.str.startswith('+57')],
          tel_fake_news_n[tel_fake_news_n.index.str.startswith('+58')])

pe = tel_fake_news_n[tel_fake_news_n.index.str.startswith('+51')]
cl = tel_fake_news_n[tel_fake_news_n.index.str.startswith('+56')]
co = tel_fake_news_n[tel_fake_news_n.index.str.startswith('+57')]
vz = tel_fake_news_n[tel_fake_news_n.index.str.startswith('+58')]
ec = tel_fake_news_n[tel_fake_news_n.index.str.startswith('+593')]

y = [np.mean(country)[0] for country in [pe, cl, co, vz, ec]]

plt.figure()
plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("Avg. # of Fake News")
plt.title("All Users from Country")
plt.savefig('images/ch-misinformation/bar_fakenews_country_frequency.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

### Number of fake news, only for fake news sharers

tel_fake_news_pos_n = tel_fake_news_n[tel_fake_news_n['fake_news'] > 0]

pe = tel_fake_news_pos_n[tel_fake_news_pos_n.index.str.startswith('+51')]['fake_news'].values
cl = tel_fake_news_pos_n[tel_fake_news_pos_n.index.str.startswith('+56')]['fake_news'].values
co = tel_fake_news_pos_n[tel_fake_news_pos_n.index.str.startswith('+57')]['fake_news'].values
vz = tel_fake_news_pos_n[tel_fake_news_pos_n.index.str.startswith('+58')]['fake_news'].values
ec = tel_fake_news_pos_n[tel_fake_news_pos_n.index.str.startswith('+593')]['fake_news'].values

fig, ax = plt.subplots()
ax.violinplot([pe, cl, co, vz, ec], showmeans = True)
ax.set_xticklabels(['','PER', 'CHL', 'COL', 'VEN', 'ECU'])
plt.xlabel("Country")
plt.ylabel("# of Fake News")
plt.title("Users Who've Shared Fake News")
plt.savefig('images/ch-misinformation/violin_fakenews_country_frequency_sharers.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

# Group dynamics

df_groups['fakeNews'] = df_p.groupby('uid')['fake_news'].mean()
df_groups['fakeNews'] = df_groups['fakeNews'].fillna(0)

df_groups['fakeNews_users'] = (df_p[['tel', 'uid', 'fake_news']].groupby(['tel', 'uid']).sum() > 1).groupby('uid').mean()
df_groups['fakeNews_users'] = df_groups['fakeNews_users'].fillna(0)

df_groups.to_csv('Data/df_groups_nb4b-fakeNews.csv')

print(np.mean(df_groups['fakeNews'] == 0))
print(np.sum(df_groups['fakeNews'] == 0))
print(np.sum(df_groups['fakeNews'] <= 0.1))

df_groups[df_groups['fakeNews'] > 0]['fakeNews'].hist(bins = 20)
plt.xlabel("Message Prevalence")
plt.ylabel("# of Groups")
plt.title("Groups with Fake News")
plt.savefig('images/ch-misinformation/hist_fakenews_group_prevalence.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.scatter(df_groups[df_groups['fakeNews'] > 0]['fakeNews'], df_groups[df_groups['fakeNews'] > 0]['fakeNews_users'], alpha = 0.2)
plt.xlabel("Message Prevalence")
plt.ylabel("User Prevalence")
plt.title("Groups with Fake News")
plt.savefig('images/ch-misinformation/scatter_fakenews_group.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(scipy.stats.pearsonr(df_groups['fakeNews'], df_groups['fakeNews_users']))
print(scipy.stats.pearsonr(df_groups[df_groups['fakeNews'] > 0]['fakeNews'], df_groups[df_groups['fakeNews'] > 0]['fakeNews_users']))
# Fake news less likely to go viral?!?!

df_groups[df_groups['fakeNews'] > 0.4]
# Small groups, one 2 messages, one 24 messages

df_p[df_p['fake_news']].groupby('uid')['fake_news'].sum().hist()
plt.xlabel("# Fake News Shared")
plt.ylabel("# of Groups")
plt.title("Groups with Fake News")
plt.savefig('images/ch-misinformation/hist_fakenews_group_frequency.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Correlates

cols = ['Size', 'pVZ', 'pCO',
       'pUS', 'pPE', 'pCL', 'pEC', 'p3rdCountry', 'entropy',
       'activity', 'degree', 'hhConc',
       'gini', 'virality']

for col in cols:
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[col], df_groups['fakeNews']))
    print(scipy.stats.pearsonr(df_groups[col], df_groups['fakeNews_users']))

reg = smf.ols('fakeNews_users ~ pVZ',
              data = df_groups).fit()
print(reg.summary())

#### Of groups that share ANY fake news

for col in cols:
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[df_groups['fakeNews'] > 0][col], df_groups[df_groups['fakeNews'] > 0]['fakeNews']))
    print(scipy.stats.pearsonr(df_groups[df_groups['fakeNews'] > 0][col], df_groups[df_groups['fakeNews'] > 0]['fakeNews_users']))
    

reg = smf.ols('fakeNews ~ Size + activity + hhConc + gini + virality',
              data = df_groups[df_groups['fakeNews'] > 0]).fit()
print(reg.summary())

reg = smf.ols('fakeNews ~ Size + activity + hhConc + gini + virality',
              data = df_groups[(df_groups['fakeNews'] > 0) & (df_groups['fakeNews'] < 0.3)]).fit()
print(reg.summary())

# Grouping fake news together

fake_news_corpus = df_x[df_x['fake_news']]['textlower'].unique()
fake_news_tokenize = [tokenize(c) for c in fake_news_corpus]

print(len(fake_news_corpus))

X = vectorizer.fit_transform(fake_news_tokenize)

redirect_from_to = {}

for i, j in zip(np.where(cosine_similarity(X) > 0.8)[0], np.where(cosine_similarity(X) > 0.8)[1]):
    if i < j:
        if i in redirect_from_to:
            redirect_from_to[j] = redirect_from_to[i]
        else:
            redirect_from_to[j] = i

print(len(redirect_from_to))

print(redirect_from_to)

print(df_p[df_p['textlower'] == fake_news_corpus[116]]['text'].iloc[0])
print()
print(df_p[df_p['textlower'] == fake_news_corpus[103]]['text'].iloc[0])

print(df_p[df_p['textlower'] == fake_news_corpus[97]]['text'].iloc[0])
print()
print(df_p[df_p['textlower'] == fake_news_corpus[67]]['text'].iloc[0])


groups_where_shared = {}
users_who_shared = {}

for i in range(len(fake_news_corpus)):
    groups_where_shared[i] = df_x.loc[df_x['textlower'] == fake_news_corpus[i], 'uid'].tolist()
    users_who_shared[i] = df_x.loc[df_x['textlower'] == fake_news_corpus[i], 'tel'].tolist()

for from_i in redirect_from_to.keys():
    to_i = redirect_from_to[from_i]
    groups_where_shared[to_i] = groups_where_shared[to_i] + groups_where_shared[from_i]
    users_who_shared[to_i] = users_who_shared[to_i] + users_who_shared[from_i]
    
    del groups_where_shared[from_i]
    del users_who_shared[from_i]

list_mergedFakeNews = \
    [[fake_news_corpus[i], groups_where_shared[i], users_who_shared[i]] for i in groups_where_shared.keys()]
df_mergedFakeNews = pd.DataFrame(list_mergedFakeNews, columns = ['textlower', 'groups', 'users'])

df_mergedFakeNews['n_shares'] = df_mergedFakeNews['groups'].apply(len)
df_mergedFakeNews['n_groups'] = df_mergedFakeNews['groups'].apply(lambda x: len(set(x)))
df_mergedFakeNews['n_users'] = df_mergedFakeNews['users'].apply(lambda x: len(set(x)))
df_mergedFakeNews['shares/groups'] = df_mergedFakeNews['n_shares'] / df_mergedFakeNews['n_groups']
df_mergedFakeNews['shares/users'] = df_mergedFakeNews['n_shares'] / df_mergedFakeNews['n_users']

df_mergedFakeNews['n_shares'].hist()
plt.xlabel("Number of Shares")
plt.ylabel("# of Fake News")
plt.title("Unique Pieces of Fake News")
plt.savefig('images/ch-misinformation/hist_fakenews_unique_frequency.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

df_mergedFakeNews['n_users'].hist()
plt.xlabel("# of Unique Users Shared By")
plt.ylabel("# of Fake News")
plt.title("Unique Pieces of Fake News")
plt.savefig('images/ch-misinformation/hist_fakenews_unique_user.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

df_mergedFakeNews['n_groups'].hist()
plt.xlabel("# of Unique Groups Shared In")
plt.ylabel("# of Fake News")
plt.title("Unique Pieces of Fake News")
plt.savefig('images/ch-misinformation/hist_fakenews_unique_group.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(np.mean(df_mergedFakeNews['shares/groups']))
print(np.mean(df_mergedFakeNews['shares/users']))

Counter(redirect_from_to.values())

print(fake_news_corpus[0])
print()
print(fake_news_corpus[124])
