==> analysis/4c_misinformation_overview-scams_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

import statsmodels.formula.api as smf
import scipy.stats
from scipy.stats import ttest_ind
from scipy.stats import f_oneway

df_x = pd.read_csv("Data/df_x_nb4a-mis.csv", index_col = 0)
print(df_x.shape)

df_x['text'] = df_x['text'].fillna('')
df_x['textlower'] = df_x['textlower'].fillna('')
df_x['virality'] = df_x['virality'].fillna(0)

### Text similarity imports/functions

import string, re, unidecode
from sklearn.decomposition import LatentDirichletAllocation as LDA

import nltk
from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer

from collections import Counter

from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity

stemmer = SnowballStemmer('spanish')
remove_punc = str.maketrans(string.punctuation, len(string.punctuation) * " ")
stopwords_ascii = [unidecode.unidecode(w) for w in stopwords.words('spanish')]

def tokenize(s):
    s = unidecode.unidecode(s)
    s = s.translate(remove_punc)
    tokens = nltk.word_tokenize(s)
    filtered = [w for w in tokens if w not in stopwords_ascii]
    return [stemmer.stem(w) for w in filtered]

def dummy(x):
    return x

vectorizer = TfidfVectorizer(analyzer='word', tokenizer=dummy, preprocessor=dummy,
                            token_pattern=None)

count_vectorizer = CountVectorizer(analyzer='word', tokenizer=dummy, preprocessor=dummy,
                            token_pattern=None)

df_groups = pd.read_csv('Data/df_groups_nb4b-fakeNews.csv', index_col = 0)
print(df_groups.head())

# Message dynamics

df_x['posMisinfo'] = (df_x['textlower'].apply(lambda x: len(tokenize(x)) >= 5))
df_p = df_x[df_x['posMisinfo']]

print(df_p[df_p['scam']].shape)
print(df_p.shape)

### Replies

reg = smf.ols('replies_n ~ scam', data = df_p).fit()
print(reg.summary())
reg = smf.ols('replies_n ~ scam', data = df_x).fit()
print(reg.summary())

print(np.quantile(df_p[df_p['scam']]['replies_n'], .95))
print(np.quantile(df_p[~df_p['scam']]['replies_n'], .95))

### Virality

reg = smf.ols('virality ~ scam', data = df_p[df_p['virality'] > 0]).fit()
print(reg.summary())
reg = smf.ols('virality ~ scam', data = df_x[df_x['virality'] > 0]).fit()
print(reg.summary())

print(np.quantile(df_p[df_p['scam'] & (df_p['virality'] > 0)]['virality'], .95))
print(np.quantile(df_p[~df_p['scam'] & (df_p['virality'] > 0)]['virality'], .95))

### Size

df_p['charlength'] = df_p['text'].apply(len)
df_p['wordlength'] = df_p['text'].apply(lambda x: x.count(" "))

print(scipy.stats.pearsonr(df_p['scam'], df_p['charlength']))
print(scipy.stats.pearsonr(df_p['scam'], df_p['wordlength']))
# Fake news less likely to go viral?!?!

reg = smf.ols('charlength ~ scam', data = df_p).fit()
print(reg.summary())
reg = smf.ols('wordlength ~ scam', data = df_p).fit()
print(reg.summary())

### LDA

df_t = df_p[df_p['scam']]
df_t.loc[:,'token'] = df_t.loc[:,'textlower'].apply(tokenize)

X = count_vectorizer.fit_transform(df_t['token'])

# Tweak the two parameters below
number_topics = 10
number_words = 10

# Create and fit the LDA model
lda = LDA(n_components=number_topics, n_jobs=-1)
lda.fit(X)

# Helper function (sourced from somewhere online)
def print_topics(model, count_vectorizer, n_top_words):
    words = count_vectorizer.get_feature_names()
    for topic_idx, topic in enumerate(model.components_):
        print("\nTopic %d:" % topic_idx)
        print(" ".join([words[i]
                        for i in topic.argsort()[:-n_top_words - 1:-1]]))
        
# Print the topics found by the LDA model
print_topics(lda, count_vectorizer, number_words)

# User dynamics

df_x[df_x['scam']]['tel'].value_counts().hist(bins = 40)
plt.xlabel("Number of Scams Shared")
plt.ylabel("# of Users")
plt.savefig('images/ch-misinformation/hist_scam_user_frequency.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(df_x[df_x['scam']]['tel'].unique().shape)

df_x[df_x['scam']]['tel'].value_counts().hist(range = (10, 60))
plt.xlabel("Number of Scams Shared")
plt.ylabel("# of Users")
plt.savefig('images/ch-misinformation/hist_scam_user_frequency_extreme.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(np.mean(df_x[df_x['scam']]['tel'].value_counts() == 1))
print(np.mean(df_x[df_x['scam']]['tel'].value_counts() == 2))

tel_scam = df_p[['tel', 'scam']].groupby('tel').mean()
tel_scam_n = df_p[['tel', 'scam']].groupby('tel').sum()

print(np.mean(tel_scam == 0))
print(tel_scam.shape)

tel_scam[tel_scam > 0].hist(bins = 20)
plt.xlabel("Prevalence of Scams")
plt.ylabel("# of Users")
plt.title(None)
plt.savefig('images/ch-misinformation/hist_scam_user_prevalence.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

(tel_scam[tel_scam > 0] == 1).sum()

df_user = df_x[['tel', 'fake_news']].groupby('tel').mean()
df_user['scam'] = df_x[['tel', 'scam']].groupby('tel').mean()

print(scipy.stats.pearsonr(df_user['fake_news'], df_user['scam']))

df_user['1_fake_news'] = df_user['fake_news'] > 0
df_user['1_scam'] = df_user['scam'] > 0

print(scipy.stats.pearsonr(df_user['1_fake_news'], df_user['1_scam']))

print(df_user[df_user['1_fake_news']]['1_scam'].mean())
print(df_user[~df_user['1_fake_news']]['1_scam'].mean())
print()
print(df_user[df_user['1_fake_news']]['scam'].mean())
print(df_user[~df_user['1_fake_news']]['scam'].mean())

print(df_user[df_user['1_scam']]['1_fake_news'].mean())
print(df_user[~df_user['1_scam']]['1_fake_news'].mean())
print()
print(df_user[df_user['1_scam']]['fake_news'].mean())
print(df_user[~df_user['1_scam']]['fake_news'].mean())

## Comparing by country

### User shares scams

x = ['PER', 'CHL', 'COL', 'VEN', 'ECU']

pe = tel_scam[tel_scam.index.str.startswith('+51')] > 0
cl = tel_scam[tel_scam.index.str.startswith('+56')] > 0
co = tel_scam[tel_scam.index.str.startswith('+57')] > 0
vz = tel_scam[tel_scam.index.str.startswith('+58')] > 0
ec = tel_scam[tel_scam.index.str.startswith('+593')] > 0

y = [100 * np.mean(country)[0] for country in [pe, cl, co, vz, ec]]

plt.figure()
plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("% of Users Who've\nShared Scams")
plt.savefig('images/ch-misinformation/bar_scam_country_whoshared.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

print(np.mean(co))
print(np.mean(vz))

### Prop of user content is scams

pe = tel_scam[tel_scam.index.str.startswith('+51')]
cl = tel_scam[tel_scam.index.str.startswith('+56')]
co = tel_scam[tel_scam.index.str.startswith('+57')]
vz = tel_scam[tel_scam.index.str.startswith('+58')]
ec = tel_scam[tel_scam.index.str.startswith('+593')]

y = [np.mean(country)[0] for country in [pe, cl, co, vz, ec]]

plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("Avg. Scam Prevalence")
plt.title("All Users from Country")
plt.savefig('images/ch-misinformation/bar_scam_country_prevalence.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

print(np.mean(co))
print(np.mean(vz))

### Prop user content is scams, ONLY FOR scams sharers

tel_scam_pos = tel_scam[tel_scam['scam'] > 0]

pe = tel_scam_pos[tel_scam_pos.index.str.startswith('+51')]['scam'].values
cl = tel_scam_pos[tel_scam_pos.index.str.startswith('+56')]['scam'].values
co = tel_scam_pos[tel_scam_pos.index.str.startswith('+57')]['scam'].values
vz = tel_scam_pos[tel_scam_pos.index.str.startswith('+58')]['scam'].values
ec = tel_scam_pos[tel_scam_pos.index.str.startswith('+593')]['scam'].values

fig, ax = plt.subplots()
ax.violinplot([pe, cl, co, vz, ec], showmeans = True)
ax.set_xticklabels(['','PER', 'CHL', 'COL', 'VEN', 'ECU'])
plt.xlabel("Country")
plt.ylabel("Scam Prevalence")
plt.title("Users Who've Shared Scams")
plt.savefig('images/ch-misinformation/violin_scam_country_prevalence_sharers.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

print(np.mean(co))
print(np.mean(vz))

### Number of scams

pe = tel_scam_n[tel_scam_n.index.str.startswith('+51')]
cl = tel_scam_n[tel_scam_n.index.str.startswith('+56')]
co = tel_scam_n[tel_scam_n.index.str.startswith('+57')]
vz = tel_scam_n[tel_scam_n.index.str.startswith('+58')]
ec = tel_scam_n[tel_scam_n.index.str.startswith('+593')]

y = [np.mean(country)[0] for country in [pe, cl, co, vz, ec]]

plt.bar(x, y)
plt.xlabel("Country")
plt.ylabel("Avg. # of Scams")
plt.title("All Users from Country")
plt.savefig('images/ch-misinformation/bar_scam_country_frequency.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

print(np.mean(co))
print(np.mean(vz))

### Number of fake news, only for fake news sharers

tel_scam_pos_n = tel_scam_n[tel_scam_n['scam'] > 0]

pe = tel_scam_pos_n[tel_scam_pos_n.index.str.startswith('+51')]['scam'].values
cl = tel_scam_pos_n[tel_scam_pos_n.index.str.startswith('+56')]['scam'].values
co = tel_scam_pos_n[tel_scam_pos_n.index.str.startswith('+57')]['scam'].values
vz = tel_scam_pos_n[tel_scam_pos_n.index.str.startswith('+58')]['scam'].values
ec = tel_scam_pos_n[tel_scam_pos_n.index.str.startswith('+593')]['scam'].values

fig, ax = plt.subplots()
ax.violinplot([pe, cl, co, vz, ec], showmeans = True)
ax.set_xticklabels(['','PER', 'CHL', 'COL', 'VEN', 'ECU'])
plt.xlabel("Country")
plt.ylabel("# of Scams")
plt.title("Users Who've Shared Scams")
plt.savefig('images/ch-misinformation/violin_scam_country_frequency_sharers.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
print(f_oneway(pe, cl, co, vz, ec))

print("\nVZ t-tests")
print(ttest_ind(vz, pe))
print(ttest_ind(vz, cl))
print(ttest_ind(vz, co))
print(ttest_ind(vz, ec))

print("\nCO t-tests")
print(ttest_ind(co, pe))
print(ttest_ind(co, cl))
print(ttest_ind(co, vz))
print(ttest_ind(co, ec))

print(np.mean(co))
print(np.mean(vz))

# Group dynamics

df_groups['scam'] = df_p.groupby('uid')['scam'].mean()
df_groups['scam'] = df_groups['scam'].fillna(0)

df_groups['scam_users'] = (df_p[['tel', 'uid', 'scam']].groupby(['tel', 'uid']).sum() > 1).groupby('uid').mean()
df_groups['scam_users'] = df_groups['scam_users'].fillna(0)

df_groups.to_csv('Data/df_groups_nb4c-scam.csv')

print(scipy.stats.pearsonr(df_groups['fakeNews'], df_groups['scam']))
print(scipy.stats.pearsonr(df_groups['fakeNews_users'], df_groups['scam_users']))

print(np.mean(df_groups['scam'] == 0))
print(np.sum(df_groups['scam'] == 0))
print(np.sum(df_groups['scam'] <= 0.1))

df_groups[df_groups['scam'] > 0]['scam'].hist(bins = 20)
plt.xlabel("Message Prevalence")
plt.ylabel("# of Groups")
plt.title("Groups with Scams")
plt.savefig('images/ch-misinformation/hist_scam_group_prevalence.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

plt.scatter(df_groups[df_groups['scam'] > 0]['scam'], df_groups[df_groups['scam'] > 0]['scam_users'], alpha = 0.2)
plt.xlabel("Message Prevalence")
plt.ylabel("User Prevalence")
plt.title("Groups with Scams")
plt.savefig('images/ch-misinformation/scatter_scam_group.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

df_groups[df_groups['scam'] >= 0.5]
# Vzs en Bogota, few messages; screwing around, few; 
# business group, few; porn group, moderate; screwing around, few;
# internet money-making, MANY; VZ general, small

df_p[df_p['scam']].groupby('uid')['scam'].sum().hist()
plt.xlabel("# of Scams Shared")
plt.ylabel("# of Groups")
plt.title("Groups with Scams")
plt.savefig('images/ch-misinformation/hist_scam_group_frequency.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()
# Very active gaming group, 156 scams shared of 1751 messages

### Correlates

cols = ['Size', 'pVZ', 'pCO',
       'pUS', 'pPE', 'pCL', 'pEC', 'p3rdCountry', 'entropy',
       'activity', 'degree', 'hhConc',
       'gini', 'virality', 'fakeNews', 'fakeNews_users']

for col in cols:
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[col], df_groups['scam']))
    print(scipy.stats.pearsonr(df_groups[col], df_groups['scam_users']))

reg = smf.ols('scam ~ entropy + virality',
              data = df_groups).fit()
print(reg.summary())

reg = smf.ols('scam_users ~ pVZ + gini',
              data = df_groups).fit()
print(reg.summary())

#### Of groups that share scams

for col in cols:
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[df_groups['scam'] > 0][col], df_groups[df_groups['scam'] > 0]['scam']))
    print(scipy.stats.pearsonr(df_groups[df_groups['scam'] > 0][col], df_groups[df_groups['scam'] > 0]['scam_users']))

reg = smf.ols('scam ~ Size + activity + degree + hhConc + gini + virality',
              data = df_groups[df_groups['scam'] > 0]).fit()
print(reg.summary())

# Grouping fake news together

scam_corpus = df_x[df_x['scam']]['textlower'].unique()
scam_tokenize = [tokenize(c) for c in scam_corpus]

print(len(scam_corpus))

X = vectorizer.fit_transform(scam_tokenize)

redirect_from_to = {}

for i, j in zip(np.where(cosine_similarity(X) > 0.8)[0], np.where(cosine_similarity(X) > 0.8)[1]):
    if i < j:
        if i in redirect_from_to:
            redirect_from_to[j] = redirect_from_to[i]
        else:
            redirect_from_to[j] = i

print(len(redirect_from_to))

print(redirect_from_to)

print(df_p[df_p['textlower'] == scam_corpus[229]]['text'].iloc[0])
print()
print(df_p[df_p['textlower'] == scam_corpus[171]]['text'].iloc[0])

groups_where_shared = {}
users_who_shared = {}

for i in range(len(scam_corpus)):
    groups_where_shared[i] = df_x.loc[df_x['textlower'] == scam_corpus[i], 'uid'].tolist()
    users_who_shared[i] = df_x.loc[df_x['textlower'] == scam_corpus[i], 'tel'].tolist()

for from_i in redirect_from_to.keys():
    to_i = redirect_from_to[from_i]
    groups_where_shared[to_i] = groups_where_shared[to_i] + groups_where_shared[from_i]
    users_who_shared[to_i] = users_who_shared[to_i] + users_who_shared[from_i]
    
    del groups_where_shared[from_i]
    del users_who_shared[from_i]

list_mergedScam = \
    [[scam_corpus[i], groups_where_shared[i], users_who_shared[i]] for i in groups_where_shared.keys()]
df_mergedScam = pd.DataFrame(list_mergedScam, columns = ['textlower', 'groups', 'users'])

df_mergedScam['n_shares'] = df_mergedScam['groups'].apply(len)
df_mergedScam['n_groups'] = df_mergedScam['groups'].apply(lambda x: len(set(x)))
df_mergedScam['n_users'] = df_mergedScam['users'].apply(lambda x: len(set(x)))
df_mergedScam['shares/groups'] = df_mergedScam['n_shares'] / df_mergedScam['n_groups']
df_mergedScam['shares/users'] = df_mergedScam['n_shares'] / df_mergedScam['n_users']

df_mergedScam['n_shares'].hist(bins = 40)
plt.xlabel("Number of Shares")
plt.ylabel("# of Scams")
plt.title("Unique Scams")
plt.savefig('images/ch-misinformation/hist_scam_unique_frequency.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

df_mergedScam['n_users'].hist(bins = 40)
plt.xlabel("# of Unique Users Sharing")
plt.ylabel("# of Scam Texts")
plt.title("Unique Scams")
plt.savefig('images/ch-misinformation/hist_scam_unique_user.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

df_mergedScam['n_groups'].hist(bins = 20)
plt.xlabel("# of Unique Groups Shared In")
plt.ylabel("# of Scam Texts")
plt.title("Unique Scams")
plt.savefig('images/ch-misinformation/hist_scam_unique_group.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(np.mean(df_mergedScam['shares/groups']))
print(np.mean(df_mergedScam['shares/users']))

Counter(redirect_from_to.values())

print(df_p[df_p['textlower'] == scam_corpus[145]]['text'].iloc[0])
print()
print(df_p[df_p['textlower'] == scam_corpus[76]]['text'].iloc[0])
