==> analysis/4d_misinformation_detecting_PRODUCTION.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta

import matplotlib.pyplot as plt
SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

from sklearn.metrics import confusion_matrix
from sklearn import tree
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB

from scipy.sparse import hstack

df_x = pd.read_csv("Data/df_x_nb4a-mis.csv", index_col = 0)
print(df_x.shape)

df_x['textlower'] = df_x['textlower'].fillna('')
df_x['message_dt'] = pd.to_datetime(df_x["message_dt"], format='%Y-%m-%d %H:%M:00')

### Text similarity imports/functions

import string, re, unidecode

import nltk
from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer

from collections import Counter

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

stemmer = SnowballStemmer('spanish')
remove_punc = str.maketrans(string.punctuation, len(string.punctuation) * " ")
stopwords_ascii = [unidecode.unidecode(w) for w in stopwords.words('spanish')]

def tokenize(s):
    s = unidecode.unidecode(s)
    s = s.translate(remove_punc)
    tokens = nltk.word_tokenize(s)
    filtered = [w for w in tokens if w not in stopwords_ascii]
    return [stemmer.stem(w) for w in filtered]

def dummy(x):
    return x

vectorizer = TfidfVectorizer(analyzer='word', tokenizer=dummy, preprocessor=dummy,
                            token_pattern=None)

scaler = StandardScaler()

# Pre-processing

df_x['token'] = df_x['textlower'].apply(tokenize)

df_x['token_length'] = df_x['token'].apply(len)
df_x = df_x[df_x['token_length'] >= 5]

print(df_x.iloc[int(df_x.shape[0] * 0.8)]['message_dt'])

df_test = df_x[df_x['message_dt'] > pd.Timestamp(2020, 3, 28, 13, 48)]
df_train = df_x[df_x['message_dt'] <= pd.Timestamp(2020, 3, 28, 13, 48)]
print(df_test.shape)
print(df_train.shape)

# Properties of scams

# CV

print(df_train.iloc[int(df_train.shape[0] * 0.2)]['message_dt'])
print(df_train.iloc[int(df_train.shape[0] * 0.4)]['message_dt'])
print(df_train.iloc[int(df_train.shape[0] * 0.6)]['message_dt'])
print(df_train.iloc[int(df_train.shape[0] * 0.8)]['message_dt'])

# Returns (recall, precision)
def rec_pre(trn, tst, clf):
    X_trn = vectorizer.fit_transform(trn['token'])
    X_tst = vectorizer.transform(tst['token'])

    clf.fit(X_trn, trn['scam'])

    tn, fp, fn, tp = confusion_matrix(tst['scam'], clf.predict(X_tst)).ravel()
    return tp / (tp + fn), tp / (tp + fp)

# Returns recall list, precision list

def forward_chain(clf, fn = rec_pre):
    stat = []
    
    # Using 0-20% of training, testing on 20-40%
    trn = df_train[df_train['message_dt'] <= pd.Timestamp(2020, 2, 22, 10)]
    tst = df_train[(df_train['message_dt'] > pd.Timestamp(2020, 2, 22, 10)) \
                      & (df_train['message_dt'] <= pd.Timestamp(2020, 3, 3, 13, 21))]
    
    stat.append(fn(trn, tst, clf))
    
    # Using 0-40% of training, testing on 40-60%
    trn = df_train[df_train['message_dt'] <= pd.Timestamp(2020, 3, 3, 13, 21)]
    tst = df_train[(df_train['message_dt'] > pd.Timestamp(2020, 3, 3, 13, 21)) \
                      & (df_train['message_dt'] <= pd.Timestamp(2020, 3, 13, 11, 47))]

    stat.append(fn(trn, tst, clf))
    
    # Using 0-60% of training, testing on 60-80%
    trn = df_train[df_train['message_dt'] <= pd.Timestamp(2020, 3, 13, 11, 47)]
    tst = df_train[(df_train['message_dt'] > pd.Timestamp(2020, 3, 13, 11, 47)) \
                      & (df_train['message_dt'] <= pd.Timestamp(2020, 3, 21, 11, 57))]
        
    stat.append(fn(trn, tst, clf))
    
    # Using 0-80% of training, testing on 80-100%
    trn = df_train[df_train['message_dt'] <= pd.Timestamp(2020, 3, 21, 11, 57)]
    tst = df_train[df_train['message_dt'] > pd.Timestamp(2020, 3, 21, 11, 57)]
        
    stat.append(fn(trn, tst, clf))
    
    return [s[0] for s in stat], [s[1] for s in stat]

# Tokens Only

### Logistic regression

#Too strong regularization results in not PREDICTING any positives. Specifically:

for c in [0.5, 0.6, 0.7, 0.8, 0.9]:
    clf = LogisticRegression(C = c)
    fc = forward_chain(clf)
    print(c, fc[0], fc[1])

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.7, 0.85]:
    print("%s..." % c, end = '')
    clf = LogisticRegression(C = c)
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 5):
    c = 10**i
    print("%s..." % i, end = '')
    clf = LogisticRegression(C = c)
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Logistic Regression with Varying C ($L_2$ Regularization)")
plt.savefig('images/ch-misinformation/ml_log.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.7, 0.85]:
    print("%s..." % c, end = '')
    clf = LogisticRegression(penalty = 'l1', C = c, solver = 'liblinear')
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 5):
    c = 10**i
    print("%s..." % i, end = '')
    clf = LogisticRegression(penalty = 'l1', C = c, solver = 'liblinear')
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Logistic Regression with Varying C ($L_1$ Regularization)")
plt.savefig('images/ch-misinformation/ml_log_l1.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### SVM

#Too strong regularization results in not PREDICTING any positives. Specifically:

for i in range(-5, 2):
    c = 10**i
    clf = SVC(C = c)
    fc = forward_chain(clf)
    print(i, fc[0], fc[1])

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.25, 0.5, 0.75]:
    print("%s..." % c, end = '')
    clf = SVC(C = c)
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 3):
    c = 10**i
    print("%s..." % i, end = '')
    clf = SVC(C = c)
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("SVM with Varying C")
plt.savefig('images/ch-misinformation/ml_svm.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Decision trees

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for md in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % md, end = '')
    clf = tree.DecisionTreeClassifier(max_depth = md)
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = md)
    ax2.plot([1, 2, 3, 4], pre, label = md)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Decision Trees of Varying Depths")
plt.savefig('images/ch-misinformation/ml_dt.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### RF

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for md in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % md, end = '')
    clf = RandomForestClassifier(max_depth = md)
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = md)
    ax2.plot([1, 2, 3, 4], pre, label = md)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Random Forest Classifier of Varying Depths")
plt.savefig('images/ch-misinformation/ml_rf.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Neighbors

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for kn in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % kn, end = '')
    clf = KNeighborsClassifier(n_neighbors = kn)
    rec, pre = forward_chain(clf)
    ax1.plot([1, 2, 3, 4], rec, label = kn)
    ax2.plot([1, 2, 3, 4], pre, label = kn)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("K-Nearest Neighbors Classifier")
plt.savefig('images/ch-misinformation/ml_knn.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# Using word length

df_x['wordlength'] = df_x['text'].apply(lambda x: x.count(" "))

# Returns (recall, precision)
def rec_pre_wl(trn, tst, clf):
    X_trn = vectorizer.fit_transform(trn['token'])
    X_tst = vectorizer.transform(tst['token'])
    
    wl_trn = scaler.fit_transform(trn['text'].apply(lambda x: x.count(" ")).values.reshape(-1, 1))
    wl_tst = scaler.fit_transform(tst['text'].apply(lambda x: x.count(" ")).values.reshape(-1, 1))
    
    X_trn = hstack([X_trn, wl_trn])
    X_tst = hstack([X_tst, wl_tst])

    clf.fit(X_trn, trn['scam'])

    tn, fp, fn, tp = confusion_matrix(tst['scam'], clf.predict(X_tst)).ravel()
    return tp / (tp + fn), tp / (tp + fp)

### Logistic regression

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.7, 0.85]:
    print("%s..." % c, end = '')
    clf = LogisticRegression(C = c)
    rec, pre = forward_chain(clf, rec_pre_wl)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 5):
    c = 10**i
    print("%s..." % i, end = '')
    clf = LogisticRegression(C = c)
    rec, pre = forward_chain(clf, rec_pre_wl)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Logistic Regression with Varying C ($L_2$ Regularization)\nUsing Tokens + # of Words")
plt.savefig('images/ch-misinformation/ml_log_wl.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### SVM

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.25, 0.5, 0.75]:
    print("%s..." % c, end = '')
    clf = SVC(C = c)
    rec, pre = forward_chain(clf, rec_pre_wl)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 3):
    c = 10**i
    print("%s..." % i, end = '')
    clf = SVC(C = c)
    rec, pre = forward_chain(clf, rec_pre_wl)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("SVM with Varying C\nUsing Tokens + # of Words")
plt.savefig('images/ch-misinformation/ml_svm_wl.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Decision trees

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for md in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % md, end = '')
    clf = tree.DecisionTreeClassifier(max_depth = md)
    rec, pre = forward_chain(clf, rec_pre_wl)
    ax1.plot([1, 2, 3, 4], rec, label = md)
    ax2.plot([1, 2, 3, 4], pre, label = md)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Decision Trees of Varying Depths\nUsing Tokens + # of Words")
plt.savefig('images/ch-misinformation/ml_dt_wl.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Neighbors

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for kn in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % kn, end = '')
    clf = KNeighborsClassifier(n_neighbors = kn)
    rec, pre = forward_chain(clf, rec_pre_wl)
    ax1.plot([1, 2, 3, 4], rec, label = kn)
    ax2.plot([1, 2, 3, 4], pre, label = kn)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("K-Nearest Neighbors Classifier\nUsing Tokens + # of Words")
plt.savefig('images/ch-misinformation/ml_knn_wl.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# Using user country code

# Returns (recall, precision)
def rec_pre_cc(trn, tst, clf):
    X_trn = vectorizer.fit_transform(trn['token'])
    X_tst = vectorizer.transform(tst['token'])
    
    co_trn = trn['tel'].str.startswith('+57 ').values.reshape(-1, 1)
    vz_trn = trn['tel'].str.startswith('+58 ').values.reshape(-1, 1)
    
    co_tst = tst['tel'].str.startswith('+57 ').values.reshape(-1, 1)
    vz_tst = tst['tel'].str.startswith('+58 ').values.reshape(-1, 1)
    
    X_trn = hstack([X_trn, co_trn, vz_trn])
    X_tst = hstack([X_tst, co_tst, vz_tst])

    clf.fit(X_trn, trn['scam'])

    tn, fp, fn, tp = confusion_matrix(tst['scam'], clf.predict(X_tst)).ravel()
    return tp / (tp + fn), tp / (tp + fp)

### Logistic regression

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.7, 0.85]:
    print("%s..." % c, end = '')
    clf = LogisticRegression(C = c)
    rec, pre = forward_chain(clf, rec_pre_cc)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 5):
    c = 10**i
    print("%s..." % i, end = '')
    clf = LogisticRegression(C = c)
    rec, pre = forward_chain(clf, rec_pre_cc)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Logistic Regression with Varying C ($L_2$ Regularization)\nUsing Tokens + User Country Code")
plt.savefig('images/ch-misinformation/ml_log_cc.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### SVM

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.25, 0.5, 0.75]:
    print("%s..." % c, end = '')
    clf = SVC(C = c)
    rec, pre = forward_chain(clf, rec_pre_cc)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 3):
    c = 10**i
    print("%s..." % i, end = '')
    clf = SVC(C = c)
    rec, pre = forward_chain(clf, rec_pre_cc)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("SVM with Varying C\nUsing Tokens + User Country Code")
plt.savefig('images/ch-misinformation/ml_svm_cc.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Decision trees

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for md in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % md, end = '')
    clf = tree.DecisionTreeClassifier(max_depth = md)
    rec, pre = forward_chain(clf, rec_pre_cc)
    ax1.plot([1, 2, 3, 4], rec, label = md)
    ax2.plot([1, 2, 3, 4], pre, label = md)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Decision Trees of Varying Depths\nUsing Tokens + User Country Code")
plt.savefig('images/ch-misinformation/ml_dt_cc.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Neighbors

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for kn in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % kn, end = '')
    clf = KNeighborsClassifier(n_neighbors = kn)
    rec, pre = forward_chain(clf, rec_pre_cc)
    ax1.plot([1, 2, 3, 4], rec, label = kn)
    ax2.plot([1, 2, 3, 4], pre, label = kn)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("K-Nearest Neighbors Classifier\nUsing Tokens + User Country Code")
plt.savefig('images/ch-misinformation/ml_knn_cc.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# Using group information

df_groups = pd.read_csv("Data/df_groups_nb3b-virality.csv", index_col = 0)
print(df_groups.shape)

group_hhConc = df_groups['hhConc'].to_dict()
group_gini = df_groups['gini'].to_dict()

df_train['hhConc'] = df_train['uid'].apply(lambda x: group_hhConc[x])
df_train['gini'] = df_train['uid'].apply(lambda x: group_gini[x])

df_test['hhConc'] = df_test['uid'].apply(lambda x: group_hhConc[x])
df_test['gini'] = df_test['uid'].apply(lambda x: group_gini[x])

# Returns (recall, precision)
def rec_pre_group(trn, tst, clf):
    X_trn = vectorizer.fit_transform(trn['token'])
    X_tst = vectorizer.transform(tst['token'])
    
    conc_trn = trn['hhConc'].values.reshape(-1, 1)
    gini_trn = trn['gini'].values.reshape(-1, 1)
    
    conc_tst = tst['hhConc'].values.reshape(-1, 1)
    gini_tst = tst['gini'].values.reshape(-1, 1)
    
    X_trn = hstack([X_trn, conc_trn, gini_trn])
    X_tst = hstack([X_tst, conc_tst, gini_tst])

    clf.fit(X_trn, trn['scam'])

    tn, fp, fn, tp = confusion_matrix(tst['scam'], clf.predict(X_tst)).ravel()
    return tp / (tp + fn), tp / (tp + fp)

### Logistic regression

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.7, 0.85]:
    print("%s..." % c, end = '')
    clf = LogisticRegression(C = c)
    rec, pre = forward_chain(clf, rec_pre_group)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 5):
    c = 10**i
    print("%s..." % i, end = '')
    clf = LogisticRegression(C = c)
    rec, pre = forward_chain(clf, rec_pre_group)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Logistic Regression with Varying C ($L_2$ Regularization)\nUsing Tokens + Group Concentration, Inequality")
plt.savefig('images/ch-misinformation/ml_log_group.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### SVM

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for c in [0.25, 0.5, 0.75]:
    print("%s..." % c, end = '')
    clf = SVC(C = c)
    rec, pre = forward_chain(clf, rec_pre_group)
    ax1.plot([1, 2, 3, 4], rec, label = c)
    ax2.plot([1, 2, 3, 4], pre, label = c)
    
for i in range(0, 3):
    c = 10**i
    print("%s..." % i, end = '')
    clf = SVC(C = c)
    rec, pre = forward_chain(clf, rec_pre_group)
    ax1.plot([1, 2, 3, 4], rec, label = "10^%s" % i)
    ax2.plot([1, 2, 3, 4], pre, label = "10^%s" % i)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("SVM with Varying C\nUsing Tokens + Group Concentration, Inequality")
plt.savefig('images/ch-misinformation/ml_svm_group.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Decision trees

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for md in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % md, end = '')
    clf = tree.DecisionTreeClassifier(max_depth = md)
    rec, pre = forward_chain(clf, rec_pre_group)
    ax1.plot([1, 2, 3, 4], rec, label = md)
    ax2.plot([1, 2, 3, 4], pre, label = md)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("Decision Trees of Varying Depths\nUsing Tokens + Group Concentration, Inequality")
plt.savefig('images/ch-misinformation/ml_dt_group.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Neighbors

fig, (ax1, ax2) = plt.subplots(1, 2, figsize = (20, 8))

for kn in [3, 6, 9, 12, 15, 18, 21, 24]:
    print("%s..." % kn, end = '')
    clf = KNeighborsClassifier(n_neighbors = kn)
    rec, pre = forward_chain(clf, rec_pre_group)
    ax1.plot([1, 2, 3, 4], rec, label = kn)
    ax2.plot([1, 2, 3, 4], pre, label = kn)
    
ax1.legend()
ax2.legend()

ax1.set_xlabel("Training Size / Validation Size")
ax2.set_xlabel("Training Size / Validation Size")

ax1.set_ylabel("Recall (Detected / Actual Scams)")
ax2.set_ylabel("Precision (Actual / Flagged Scams)")

plt.suptitle("K-Nearest Neighbors Classifier\nUsing Tokens + Group Concentration, Inequality")
plt.savefig('images/ch-misinformation/ml_knn_group.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# Test

### Nearest neighbors

X_train = vectorizer.fit_transform(df_train['token'])
X_test = vectorizer.transform(df_test['token'])

clf = KNeighborsClassifier(n_neighbors = 3)
clf.fit(X_train, df_train['scam'])

print(confusion_matrix(df_test['scam'], clf.predict(X_test)))

### Decision tree

wl_train = scaler.fit_transform(df_train['text'].apply(lambda x: x.count(" ")).values.reshape(-1, 1))
wl_test = scaler.fit_transform(df_test['text'].apply(lambda x: x.count(" ")).values.reshape(-1, 1))

co_train = df_train['tel'].str.startswith('+57 ').values.reshape(-1, 1)
vz_train = df_train['tel'].str.startswith('+58 ').values.reshape(-1, 1)

co_test = df_test['tel'].str.startswith('+57 ').values.reshape(-1, 1)
vz_test = df_test['tel'].str.startswith('+58 ').values.reshape(-1, 1)

X_train_stack = hstack([X_train, co_train, vz_train, wl_train])
X_test_stack = hstack([X_test, co_test, vz_test, wl_test])

clf = tree.DecisionTreeClassifier(max_depth = 12)
clf.fit(X_train_stack, df_train['scam'])

print(confusion_matrix(df_test['scam'], clf.predict(X_test_stack)))

wl_train = scaler.fit_transform(df_train['text'].apply(lambda x: x.count(" ")).values.reshape(-1, 1))
wl_test = scaler.fit_transform(df_test['text'].apply(lambda x: x.count(" ")).values.reshape(-1, 1))

co_train = df_train['tel'].str.startswith('+57 ').values.reshape(-1, 1)
vz_train = df_train['tel'].str.startswith('+58 ').values.reshape(-1, 1)

co_test = df_test['tel'].str.startswith('+57 ').values.reshape(-1, 1)
vz_test = df_test['tel'].str.startswith('+58 ').values.reshape(-1, 1)

X_train_stack = hstack([X_train, co_train, vz_train, wl_train])
X_test_stack = hstack([X_test, co_test, vz_test, wl_test])

clf = tree.DecisionTreeClassifier(max_depth = 3)
clf.fit(X_train_stack, df_train['scam'])

print(confusion_matrix(df_test['scam'], clf.predict(X_test_stack)))

from sklearn.tree import export_text
from sklearn.tree import export_graphviz

tree_rules = export_text(clf)
print(tree_rules)

token_names = vectorizer.get_feature_names() + ['CO User', 'VZ User', '# Words']

dotfile = open("images/ch-misinformation/tree.dot", 'w')
export_graphviz(clf, out_file=dotfile, feature_names=token_names)
dotfile.close()
