==> analysis/5_coronavirus.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta, time
import os

import matplotlib.pyplot as plt
import matplotlib.dates as mdates

SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

import statsmodels.formula.api as smf
import scipy.stats
from scipy.stats import ttest_ind
from scipy.stats import f_oneway

df_x = pd.read_csv("Data/df_x_nb4a-mis.csv", index_col = 0)
print(df_x.shape)

df_x['text'] = df_x['text'].fillna('')
df_x['textlower'] = df_x['textlower'].fillna('')
df_x['message_dt'] = pd.to_datetime(df_x["message_dt"], format='%Y-%m-%d %H:%M:00')
df_x['message_date'] = pd.to_datetime(df_x["message_date"], format='%Y-%m-%d')

## Feature dummy

df_x['coronavirus'] = df_x['textlower'].str.contains('corona') & df_x['textlower'].str.contains('virus')
df_x['virus'] = df_x['textlower'].str.contains('virus')
df_x['frontera'] = df_x['textlower'].str.contains('frontera')
df_x['troch'] = df_x['textlower'].str.contains('troch')
df_x['cuarentena'] = df_x['textlower'].str.contains('cuarentena')

# Graph over time

## Select messages 5+ words

df_x['5_word'] = df_x['textlower'].apply(lambda x: x.count(' ') >= 4)

df_y = df_x[df_x['5_word']]

coronavirus = df_y[['message_date', 'coronavirus']].groupby('message_date').mean()
virus = df_y[['message_date', 'virus']].groupby('message_date').mean()
frontera = df_y[['message_date', 'frontera']].groupby('message_date').mean()
troch = df_y[['message_date', 'troch']].groupby('message_date').mean()
cuarentena = df_y[['message_date', 'cuarentena']].groupby('message_date').mean()

fig, ax = plt.subplots(figsize = (10, 6))
ax.plot(coronavirus * 100, label = '"Coronavirus"')
ax.plot(virus * 100, label = '"Virus"')
ax.plot(cuarentena * 100, label = '"Cuarentena"')

plt.xlabel("UTC-5 / Colombia Time")
plt.ylabel("% of 5+ Word Messages")
ax.xaxis.set_major_locator(mdates.DayLocator(interval = 14))
ax.legend()
plt.xticks(horizontalalignment = 'center', rotation = '0')
plt.savefig('images/ch-coronavirus/5word_proportion.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

fig, ax = plt.subplots(figsize = (14, 6))
ax.axvline(x = pd.Timestamp(2020, 3, 13), color = 'r', lw = 3)
ax.plot(coronavirus * 100, label = '"Coronavirus"')
ax.plot(frontera * 100, label = '"Frontera"')
ax.plot(troch * 100, label = '"Troch"')

plt.xlabel("UTC-5 / Colombia Time")
plt.ylabel("% of 5+ Word Messages")
ax.xaxis.set_major_locator(mdates.DayLocator(interval = 14))
ax.legend()
plt.xticks(horizontalalignment = 'center', rotation = '0')
plt.savefig('images/ch-coronavirus/5word_proportion_border.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

print(df_y[df_y['message_date'] < pd.Timestamp(2020, 3, 13)]['troch'].mean())
print(df_y[df_y['message_date'] < pd.Timestamp(2020, 3, 13)]['frontera'].mean())


### t-test

outside = df_y[(df_y['message_date'] < pd.Timestamp(2020, 3, 13)) \
               | (df_y['message_date'] > pd.Timestamp(2020, 3, 15))]
inside = df_y[(df_y['message_date'] >= pd.Timestamp(2020, 3, 13)) \
              & (df_y['message_date'] <= pd.Timestamp(2020, 3, 15))]

print(outside['troch'].mean())
print(inside['troch'].mean())

print(ttest_ind(outside['troch'], inside['troch']))

print(outside['frontera'].mean())
print(inside['frontera'].mean())

print(ttest_ind(outside['frontera'], inside['frontera']))

# Group %

virus = df_x[['uid', 'message_date', 'virus']].groupby(['uid', 'message_date']).sum() > 0
coronavirus = df_x[['uid', 'message_date', 'coronavirus']].groupby(['uid', 'message_date']).sum() > 0
frontera = df_x[['uid', 'message_date', 'frontera']].groupby(['uid', 'message_date']).sum() > 0
troch = df_x[['uid', 'message_date', 'troch']].groupby(['uid', 'message_date']).sum() > 0
cuarentena = df_x[['uid', 'message_date', 'cuarentena']].groupby(['uid', 'message_date']).sum() > 0

virus = virus.groupby('message_date').mean()
coronavirus = coronavirus.groupby('message_date').mean()
frontera = frontera.groupby('message_date').mean()
troch = troch.groupby('message_date').mean()
cuarentena = cuarentena.groupby('message_date').mean()


fig, ax = plt.subplots(figsize = (14, 6))
ax.axvline(x = pd.Timestamp(2020, 3, 13), color = 'r', lw = 3)
ax.plot(coronavirus * 100, label = '"Coronavirus"')
ax.plot(frontera * 100, label = '"Frontera"')
ax.plot(troch * 100, label = '"Troch"')

plt.xlabel("UTC-5 / Colombia Time")
plt.ylabel("% of Active Groups with Keyword")
ax.xaxis.set_major_locator(mdates.DayLocator(interval = 14))
ax.legend()
plt.xticks(horizontalalignment = 'center', rotation = '0')
plt.savefig('images/ch-coronavirus/group_proportion.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# % of user

virus = df_x[['tel', 'message_date', 'virus']].groupby(['tel', 'message_date']).sum() > 0
coronavirus = df_x[['tel', 'message_date', 'coronavirus']].groupby(['tel', 'message_date']).sum() > 0
frontera = df_x[['tel', 'message_date', 'frontera']].groupby(['tel', 'message_date']).sum() > 0
troch = df_x[['tel', 'message_date', 'troch']].groupby(['tel', 'message_date']).sum() > 0
cuarentena = df_x[['tel', 'message_date', 'cuarentena']].groupby(['tel', 'message_date']).sum() > 0

virus = virus.groupby('message_date').mean()
coronavirus = coronavirus.groupby('message_date').mean()
frontera = frontera.groupby('message_date').mean()
troch = troch.groupby('message_date').mean()
cuarentena = cuarentena.groupby('message_date').mean()


fig, ax = plt.subplots(figsize = (14, 6))
ax.axvline(x = pd.Timestamp(2020, 3, 13), color = 'r', lw = 3)
ax.plot(coronavirus * 100, label = '"Coronavirus"')
ax.plot(frontera * 100, label = '"Frontera"')
ax.plot(troch * 100, label = '"Troch"')

plt.xlabel("UTC-5 / Colombia time")
plt.ylabel("% of Active Users\nWho Discuss Keyword")
ax.xaxis.set_major_locator(mdates.DayLocator(interval = 14))
ax.legend()
plt.xticks(horizontalalignment = 'center', rotation = '0')
plt.savefig('images/ch-coronavirus/user_proportion.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Users with meaningful text messages

virus = df_y[['tel', 'message_date', 'virus']].groupby(['tel', 'message_date']).sum() > 0
coronavirus = df_y[['tel', 'message_date', 'coronavirus']].groupby(['tel', 'message_date']).sum() > 0
frontera = df_y[['tel', 'message_date', 'frontera']].groupby(['tel', 'message_date']).sum() > 0
troch = df_y[['tel', 'message_date', 'troch']].groupby(['tel', 'message_date']).sum() > 0
cuarentena = df_y[['tel', 'message_date', 'cuarentena']].groupby(['tel', 'message_date']).sum() > 0

virus = virus.groupby('message_date').mean()
coronavirus = coronavirus.groupby('message_date').mean()
frontera = frontera.groupby('message_date').mean()
troch = troch.groupby('message_date').mean()
cuarentena = cuarentena.groupby('message_date').mean()


fig, ax = plt.subplots(figsize = (14, 6))
ax.axvline(x = pd.Timestamp(2020, 3, 13), color = 'r', lw = 3)
ax.plot(coronavirus * 100, label = '"Coronavirus"')
ax.plot(frontera * 100, label = '"Frontera"')
ax.plot(troch * 100, label = '"Troch"')

plt.xlabel("UTC-5 / Colombia time")
plt.ylabel("% of Active Users\nWho Discuss Keyword")
plt.title("Users Who Send 5+ Word Text Messages")
ax.xaxis.set_major_locator(mdates.DayLocator(interval = 14))
ax.legend()
plt.xticks(horizontalalignment = 'center', rotation = '0')
plt.savefig('images/ch-coronavirus/user_proportion_text.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

# Group regressions

df_groups = pd.read_csv('Data/df_groups_nb4c-scam.csv', index_col = 0)

group_troch_pre = df_x[(df_x['message_date'] <= pd.Timestamp(2020, 3, 12)) & df_x['troch']]['uid'].unique()
group_troch_after = df_x[(df_x['message_date'] >= pd.Timestamp(2020, 3, 13)) & df_x['troch']]['uid'].unique()

df_groups['troch_pre'] = False
df_groups.loc[group_troch_pre, 'troch_pre'] = True

df_groups['troch_after'] = 0 # SMF requires numeric
df_groups.loc[group_troch_after, 'troch_after'] = 1

print(df_groups['troch_pre'].sum())
print(df_groups['troch_after'].sum())

cols = ['Size', 'pVZ', 'pCO', 'pUS',
       'pPE', 'pCL', 'pEC', 'p3rdCountry', 'entropy', 'activity',
       'degree', 'hhConc', 'gini',
       'replies_n', 'virality', 'fakeNews', 'fakeNews_users', 'scam',
       'scam_users', 'troch_pre']

for col in cols:
    print("============= %s =============" % col)
    print(scipy.stats.pearsonr(df_groups[col], df_groups['troch_after']))

reg = smf.probit('troch_after ~ troch_pre + Size + pVZ + pCO + degree + hhConc + gini', data = df_groups).fit()
print(reg.summary())

print(reg.get_margeff().summary())
