==> analysis/5a_coronavirus_quarantine.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta, time
import os

import matplotlib.pyplot as plt
import matplotlib.dates as mdates

SMALL_SIZE = 8
MEDIUM_SIZE = 10
BIGGER_SIZE = 20

plt.rc('font', size=BIGGER_SIZE)          # controls default text sizes
plt.rc('axes', titlesize=BIGGER_SIZE)     # fontsize of the axes title
plt.rc('axes', labelsize=BIGGER_SIZE)    # fontsize of the x and y labels
plt.rc('xtick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('ytick', labelsize=BIGGER_SIZE)    # fontsize of the tick labels
plt.rc('legend', fontsize=BIGGER_SIZE)    # legend fontsize
plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title

import statsmodels.formula.api as smf
import scipy.stats
from scipy.stats import ttest_ind
from scipy.stats import f_oneway

df_x = pd.read_csv("Data/df_x_nb4a-mis.csv", index_col = 0)
print(df_x.shape)

df_x['message_dt'] = pd.to_datetime(df_x["message_dt"], format='%Y-%m-%d %H:%M:00')
df_x['message_date'] = pd.to_datetime(df_x["message_date"], format='%Y-%m-%d')

# AM Usage pattern

df_x['prequar'] = df_x['message_date'] <= pd.Timestamp(2020, 3, 10)
df_x['quar'] = df_x['message_date'] >= pd.Timestamp(2020, 3, 24)
df_x['message_satsun'] = df_x['message_date'].dt.weekday >= 5

df_x['round30m'] = df_x['message_dt'].dt.round('30T').apply(lambda x: x.time())
df_x['roundhr'] = df_x['message_dt'].dt.round('H').apply(lambda x: x.time())

### Baseline

### Graphs

x = [(pd.Timestamp(2017, 1, 1, h), pd.Timestamp(2017, 1, 1, h, 30)) for h in range(0, 24)]
x = [j for i in x for j in i]

y_pre = df_x[df_x['prequar'] & ~df_x['message_satsun'] & \
             (df_x['tel'].str.startswith('+57'))].groupby('round30m').size()
y_quar = df_x[df_x['quar'] & ~df_x['message_satsun'] & \
             (df_x['tel'].str.startswith('+57'))].groupby('round30m').size()

y_pre = y_pre / y_pre.sum()
y_quar = y_quar / y_quar.sum()

fig, ax = plt.subplots(figsize = (8, 6))
ax.plot(x, y_pre, label = 'Before 3/10')
ax.plot(x, y_quar, label = 'After 3/24')

ax.xaxis.set_major_locator(mdates.HourLocator(interval = 8))
ax.xaxis.set_major_formatter(mdates.DateFormatter('%H:%M:%S'))

fig.autofmt_xdate()
plt.xticks(ha='center', rotation=0)
plt.xlabel("UTC-5 / Colombia Time\n(30 Minute Buckets)")
plt.ylabel("Density")
plt.title("Weekday Messages (Colombia Users)")
plt.legend()
plt.savefig('images/ch-coronavirus/usage_weekday_co.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

x = [(pd.Timestamp(2017, 1, 1, h), pd.Timestamp(2017, 1, 1, h, 30)) for h in range(0, 24)]
x = [j for i in x for j in i]

y_pre = df_x[df_x['prequar'] & df_x['message_satsun'] & \
             (df_x['tel'].str.startswith('+57'))].groupby('round30m').size()
y_quar = df_x[df_x['quar'] & df_x['message_satsun'] & \
             (df_x['tel'].str.startswith('+57'))].groupby('round30m').size()

y_pre = y_pre / y_pre.sum()
y_quar = y_quar / y_quar.sum()

fig, ax = plt.subplots(figsize = (8, 6))
ax.plot(x, y_pre, label = 'Before 3/10')
ax.plot(x, y_quar, label = 'After 3/24')

ax.xaxis.set_major_locator(mdates.HourLocator(interval = 8))
ax.xaxis.set_major_formatter(mdates.DateFormatter('%H:%M:%S'))

fig.autofmt_xdate()
plt.xticks(ha='center', rotation=0)
plt.xlabel("UTC-5 / Colombia Time\n(30 Minute Buckets)")
plt.ylabel("Density")
plt.title("Weekend Messages (Colombia Users)")
plt.legend()
plt.savefig('images/ch-coronavirus/usage_weekend_co.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

x = [pd.Timestamp(2017, 1, 1, h) for h in range(0, 24)]

y_pre = df_x[df_x['prequar'] & ~df_x['message_satsun'] & \
             (df_x['tel'].str.startswith('+58'))].groupby('roundhr').size()
y_quar = df_x[df_x['quar'] & ~df_x['message_satsun'] & \
             (df_x['tel'].str.startswith('+58'))].groupby('roundhr').size()

y_pre = y_pre / y_pre.sum()
y_quar = y_quar / y_quar.sum()

fig, ax = plt.subplots(figsize = (8, 6))
ax.plot(x, y_pre, label = 'Before 3/10')
ax.plot(x, y_quar, label = 'After 3/24')

ax.xaxis.set_major_locator(mdates.HourLocator(interval = 8))
ax.xaxis.set_major_formatter(mdates.DateFormatter('%H:%M:%S'))

fig.autofmt_xdate()
plt.xticks(ha='center', rotation=0)
plt.xlabel("UTC-5 / Colombia Time\n(Hour Buckets)")
plt.ylabel("Density")
plt.title("Weekday Messages (Venezuela Users)")
plt.legend()
plt.savefig('images/ch-coronavirus/usage_weekday_vz.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

x = [pd.Timestamp(2017, 1, 1, h) for h in range(0, 24)]

y_pre = df_x[df_x['prequar'] & df_x['message_satsun'] & \
             (df_x['tel'].str.startswith('+58'))].groupby('roundhr').size()
y_quar = df_x[df_x['quar'] & df_x['message_satsun'] & \
             (df_x['tel'].str.startswith('+58'))].groupby('roundhr').size()

y_pre = y_pre / y_pre.sum()
y_quar = y_quar / y_quar.sum()

fig, ax = plt.subplots(figsize = (8, 6))
ax.plot(x, y_pre, label = 'Before 3/10')
ax.plot(x, y_quar, label = 'After 3/24')

ax.xaxis.set_major_locator(mdates.HourLocator(interval = 8))
ax.xaxis.set_major_formatter(mdates.DateFormatter('%H:%M:%S'))

fig.autofmt_xdate()
plt.xticks(ha='center', rotation=0)
plt.xlabel("UTC-5 / Colombia Time\n(Hour Buckets)")
plt.ylabel("Density")
plt.title("Weekend Messages (Venezuela Users)")
plt.legend()
plt.savefig('images/ch-coronavirus/usage_weekend_vz.png', bbox_inches = 'tight', pad_inches = 0.05)
plt.close()

### Statistical tests

df_x['5am'] = (df_x['message_dt'].dt.time <= time(5))

prequar = df_x[df_x['prequar'] & ~df_x['message_satsun'] & \
                      df_x['tel'].str.startswith('+57')][['message_date', '5am']].groupby('message_date').mean()

quar = df_x[df_x['quar'] & ~df_x['message_satsun'] & \
                      df_x['tel'].str.startswith('+57')][['message_date', '5am']].groupby('message_date').mean()

print(prequar.mean())
print(quar.mean())
print(ttest_ind(prequar, quar))

prequar = df_x[df_x['prequar'] & df_x['message_satsun'] & \
                      df_x['tel'].str.startswith('+57')][['message_date', '5am']].groupby('message_date').mean()

quar = df_x[df_x['quar'] & df_x['message_satsun'] & \
                      df_x['tel'].str.startswith('+57')][['message_date', '5am']].groupby('message_date').mean()

print(prequar.mean())
print(quar.mean())
print(ttest_ind(prequar, quar))

# Audio/video lengths

df_text = df_x[df_x['text'].notnull()]
df_audio = df_x[df_x['audio_length'].notnull()]
df_video = df_x[df_x['video_length'].notnull()]

df_text['seconds'] = (df_text['message_dt'] - pd.Timestamp(2020, 2, 13)).dt.seconds
df_audio['seconds'] = (df_audio['message_dt'] - pd.Timestamp(2020, 2, 13)).dt.seconds
df_video['seconds'] = (df_video['message_dt'] - pd.Timestamp(2020, 2, 13)).dt.seconds


def get_sec(t):
    if t.count(':') == 1:
        m, s = t.split(':')
        h = 0
    else:
        h, m, s = t.split(':')
    return int(h) * 3600 + int(m) * 60 + int(s)

df_audio['audio_length'] = df_audio['audio_length'].apply(get_sec)
df_video['video_length'] = df_video['video_length'].apply(get_sec)
df_text['word_length'] = df_text['text'].str.count(' ')
df_text['char_length'] = df_text['text'].str.len()

## All messages

prequar = df_text[df_text['prequar']]['word_length']
quar = df_text[df_text['quar']]['word_length']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(quar, prequar)

reg = smf.ols('word_length ~ seconds', data = df_text).fit()
print(reg.summary())

prequar = df_text[df_text['prequar']]['char_length']
quar = df_text[df_text['quar']]['char_length']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(quar, prequar)

reg = smf.ols('char_length ~ seconds', data = df_text).fit()
print(reg.summary())

prequar = df_audio[df_audio['prequar']]['audio_length']
quar = df_audio[df_audio['quar']]['audio_length']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(quar, prequar)

reg = smf.ols('audio_length ~ seconds', data = df_audio).fit()
print(reg.summary())

prequar = df_video[df_video['prequar']]['video_length']
quar = df_video[df_video['quar']]['video_length']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(quar, prequar)

reg = smf.ols('video_length ~ seconds', data = df_video).fit()
print(reg.summary())

# Orig

df_orig_audio = df_audio[~df_audio['forwarded'] & ~df_audio['forwarded_highly']]
df_orig_video = df_video[~df_video['forwarded'] & ~df_video['forwarded_highly']]
df_orig_text = df_text[~df_text['forwarded'] & ~df_text['forwarded_highly']]


prequar = df_orig_text[df_orig_text['prequar']]['word_length']
quar = df_orig_text[df_orig_text['quar']]['word_length']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(quar, prequar)

prequar = df_orig_text[df_orig_text['prequar']]['char_length']
quar = df_orig_text[df_orig_text['quar']]['char_length']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(quar, prequar)

prequar = df_orig_audio[df_orig_audio['prequar']]['audio_length']
quar = df_orig_audio[df_orig_audio['quar']]['audio_length']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(quar, prequar)

prequar = df_orig_video[df_orig_video['prequar']]['video_length']
quar = df_orig_video[df_orig_video['quar']]['video_length']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(quar, prequar)

reg = smf.ols('word_length ~ seconds', data = df_orig_text).fit()
print(reg.summary())

reg = smf.ols('char_length ~ seconds', data = df_orig_text).fit()
print(reg.summary())

reg = smf.ols('audio_length ~ seconds', data = df_orig_audio).fit()
print(reg.summary())

reg = smf.ols('video_length ~ seconds', data = df_orig_video).fit()
print(reg.summary())

# Messages per user

### Weight by each user/message

prequar = df_x[df_x['prequar']][['tel', 'message_date']].groupby(['tel', 'message_date']).size()
quar = df_x[df_x['quar']][['tel', 'message_date']].groupby(['tel', 'message_date']).size()

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(prequar, quar)

df_fals = pd.DataFrame(df_x[['tel', 'message_date']].groupby(['tel', 'message_date']).size())
df_fals.columns = ['n']

df_fals['date'] = [i[1] for i in df_fals.index]
df_fals['days_since'] = (df_fals['date'] - pd.Timestamp(2020, 2, 13)).dt.days

reg = smf.ols('n ~ days_since', data = df_fals).fit()
print(reg.summary())

# Replies

prequar = df_x[df_x['prequar']]['replies_n']
quar = df_x[df_x['quar']]['replies_n']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(prequar, quar)

df_x['seconds'] = (df_x['message_dt'] - pd.Timestamp(2020, 2, 13)).dt.seconds

prequar = df_x[df_x['prequar'] & (df_x['virality'] > 0)]['virality']
quar = df_x[df_x['quar'] & (df_x['virality'] > 0)]['virality']

print(np.mean(prequar))
print(np.mean(quar))
ttest_ind(prequar, quar)

reg = smf.ols('virality ~ seconds', data = df_x[df_x['virality'] > 0]).fit()
print(reg.summary())
